% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_plots.R
\name{add_sample_lines}
\alias{add_sample_lines}
\title{Sample Lines}
\usage{
add_sample_lines(object, n = 5, sample_from_prior = FALSE, ...)
}
\arguments{
\item{object}{a \code{blm} object}

\item{n}{number of lines to show}

\item{sample_from_prior}{draw samples from the prior distribution?}

\item{...}{additional arguments passed to lines}
}
\description{
Adds random sampled lines from a \code{blm} object to an existing plot.
}
\details{
Draws random samples from the posterior (or prior if sample_from_prior = TRUE) distribution of \code{\link{blm}} object. Requires the mvrnorm from the MASS package for random sampling. NOTE: only works for straight lines, assuming first and second coefficient are intercept and slope.
}
\examples{
x <- rnorm(100)
b <- 1.3
w0 <- 0.2 ; w1 <- 3

y <- rnorm(100, mean = w0 + w1 * x, sd = sqrt(1/b))
model <- blm(y ~ x, prior = NULL, beta = b, 
               data = data.frame(x=x, y=y)) 

plot(model)
add_sample_lines(model, col='green')                

}

