% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_analysis.R
\name{anova.blm}
\alias{anova.blm}
\title{F-Test}
\usage{
\method{anova}{blm}(model1, model2, ...)
}
\arguments{
\item{model1}{first, more simple, \code{blm} model, nested in model2.#}

\item{model2}{second, more complex, \code{blm} object.}

\item{...}{other arguments (currently ignored).}
}
\value{
Analysis of Variance Table as for \code{\link{anova.lm}}.
}
\description{
Anova F-test for comparison of the fits of 2 \code{blm} objects.
}
\details{
Extracts relevant features and compute the F value and the p-value
  for comparison 2 model fits.
}
\examples{
set.seed(1) 
x <- seq(-10,10,.1) 
b <- 0.3

w0 <- 0.2 ; w1 <- 3 ; w2 <- 10

y <- rnorm(201, mean = w0 + w1 * x + w2 *sin(x), sd = sqrt(1/b)) 
mod1 <- blm(y ~ x + sin(x))

\dontrun{
  plot(mod1, xlim=c(-10,10)) 
}

#another mod removing the sinus term, clearly less well fitting
mod2 <- blm(y ~ x)

anova(mod1, mod2)

##much less distinguished model
b <- 0.003
y <- rnorm(201, mean = w0 + w1 * x + w2 *sin(x), sd = sqrt(1/b)) 
mod1 <- blm(y ~ x + sin(x))
\dontrun{
  plot(mod1, xlim=c(-10,10)) 
}
mod2 <- blm(y ~ x)
anova(mod1, mod2) #... still very low p-value !?
  
}

