% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_posterior.R
\name{append_samples.posterior}
\alias{append_samples.posterior}
\title{Append samples to a posterior distribution}
\usage{
\method{append_samples}{posterior}(x, updates)
}
\description{
Append samples to a posterior distribution
}
