% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_analysis.R
\name{bayes_factor}
\alias{bayes_factor}
\title{Bayes Factor}
\usage{
bayes_factor(model1, model2)
}
\arguments{
\item{model1}{first \code{blm} object}

\item{model2}{second \code{blm} object}
}
\description{
Computes the Bayes factor for comparison of 2 \code{\link{blm}} objects.
}
\details{
Returns the ratio of the probability of model1 divided by the
  probability of model2. A value of K > 1 means that M1 is more strongly
  supported by the data under consideration than M2
  (\url{https://en.wikipedia.org/wiki/Bayes_factor}). Wikipedia mentions
  additional scales for interpretation, ie by Harold Jeffreys:
   \itemize{
     \item K < 1: supports model2
     \item K > 10: strong support for model1
     \item K > 100: decisive support for model1
   }.
 An alternative table, widely cited, is provided by Kass and Raftery.
   \itemize{
     \item K 1-3: not worth more than a bare mention
     \item K 3-20: positive support for model1
     \item K 20-150: strong support for model1
     \item K >150: very strong support for model1
   }.
}
\examples{
set.seed(1) 
x <- seq(-10,10,.1) 
b <- 0.3

w0 <- 0.2 ; w1 <- 3 ; w2 <- 10

y <- rnorm(201, mean = w0 + w1 * x + w2 *sin(x), sd = sqrt(1/b)) 
mod1 <- blm(y ~ x + sin(x))

pblm(mod1) ## -886.4801 this is the log-probability (log=TRUE by default)

#another model removing the sinus term 
mod2 <- blm(y ~ x)

pblm(mod2) ## -891.0581 
bayes_factor(mod1, mod2) #97.31074 #strong support for mod1 over mod2

mod3 <- blm(y ~ x + 0)
pblm(mod3) ## -891.842
bayes_factor(mod1, mod3) #213.128 #very strong support for mod1 over mod3

}

