% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BayesLinReg-package}
\alias{BayesLinReg}
\alias{BayesLinReg-package}
\title{BayesLinReg package}
\description{
Bayesian linear regression with normal, spike-and-slab, and global-local
coefficient priors.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
