% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{bic}
\alias{bic}
\title{Bayesian information criterion (BIC)}
\usage{
bic(object, ...)
}
\arguments{
\item{object}{a \code{blm} object.}

\item{...}{other arguments (currently ignored).}
}
\description{
Bayesian information criterion (BIC) for the fit of a \code{blm} object.
}
\details{
Bayesian information criterion (BIC) of the \code{blm} object. 
  \code{BIC} = n * ln(RSS/n) + k * ln(n). Where n is the number of
  observations, k is the degrees of freedom and RSS (\code{\link{deviance}}) 
  is the residuals sum of squares. The strength of the evidence against the
  model with the higher BIC value can be summarized as follows according to
  Wikipedia
  (\url{https://en.wikipedia.org/wiki/Bayesian_information_criterion}):
  \tabular{ll}{
    ΔBIC \tab Evidence against higher BIC\cr
    0-2 \tab Not worth more than a bare mention\cr
    2-6 \tab Positive\cr
    6-10 \tab Strong\cr
    >10 \tab Very Strong\cr
  }
}
\examples{
  x <- rnorm(10)
  b <- 1.3
  w0 <- 0.2 ; w1 <- 3; w2 <- 10
  y <- rnorm(10, mean = w0 + w1 * x + w2 * sin(x), sd = sqrt(1/b))
  model <- blm(y ~ x + sin(x), prior = NULL, beta = b, data = data.frame(x=x, y=y))
  
  bic(model)
  
  #another model removing the sinus term
  model2 <- blm(y ~ x, prior = NULL, beta = b, data = data.frame(x=x, y=y))
  
  bic(model2)
  
}

