% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_julia_setup.R
\name{blm_setup}
\alias{blm_setup}
\title{Initial setup for blm package.}
\usage{
blm_setup(...)
}
\arguments{
\item{...}{arguments passed to \code{\link{JuliaCall::julia_setup}}.}
}
\description{
\code{blm_setup} runs the julia setup for the blm package
}
\examples{
\dontrun{
blm_setup()
}

}
