% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{blmsim}
\alias{blmsim}
\title{Generate a dataset with normal variables and outcome for testing purposes}
\usage{
blmsim(n = 2000, j = 5, binary = 1, seed = NULL,
  heteroskedastic = FALSE, correlated_errors = FALSE, ...)
}
\arguments{
\item{n}{number of examples}

\item{j}{number of variables}

\item{binary}{number of columns with 0/1 outcomes}

\item{seed}{seed for pseudo-random number generator}
}
\value{
list containing true values and data frame with n rows and j columns with simulated data.
}
\description{
Generate a dataset with normal variables and outcome for testing purposes
}
