\name{coef}
\docType{methods}
\alias{coef.blm}
\alias{coef,blm-method}
\alias{coef,lexpit-method}

\title{Get coefs from \code{blm} and \code{lexpit} objects.}

\description{

Extract vector of coefs of the fit of a \code{blm} or \code{lexpit} model.

}

\section{Methods}{
\describe{

\item{coef}{\code{signature(object = "blm")}:
Extractor for MLEs returned as a matrix with one column.
  }
 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
