% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coefficients.blm}
\alias{coefficients.blm}
\title{Coefficients for the blm class}
\usage{
\method{coefficients}{blm}(x)
}
\arguments{
\item{x}{object of class blm}
}
\value{
vector of coefficients
}
\description{
Retrieve coefficients from the blm class
}

