% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coefficients}
\alias{coefficients}
\title{Bayesian linear model.}
\usage{
coefficients(x)
}
\arguments{
\item{x}{A formula describing the model.}
}
\value{
A fitted model.
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}

