% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{compute_null_model}
\alias{compute_null_model}
\title{Compute the intercept-only model for comparison purposes}
\usage{
compute_null_model(x, iterations = 10000, chains = 1, thinning = 1,
  burn = 1000)
}
\arguments{
\item{x}{blm object}

\item{iterations}{number of iterations used to sample the model}

\item{chains}{number of separate chains used to sample the model}

\item{thinning}{thinning parameter}

\item{burn}{number of examples to burn}
}
\value{
blm object containing instructions on how to sample the null model
}
\description{
Compute the intercept-only model for comparison purposes
}
\examples{
data("directors")
fit <- blm("Compensation ~ Age", data=directors) \%>\%
   compute_null_model(iterations=10000, burn=1000) \%>\%
   sample_posterior()
# Plot the null model
plot(fit, "nullmodel")

}
