% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confidenceInterval}
\alias{confidenceInterval}
\title{Confidence Interval for the fitted parameters}
\usage{
confidenceInterval(object, level = 0.05)
}
\arguments{
\item{object}{A fiited model obtained from the blm}

\item{level}{A confidence level}
}
\value{
Confidence Interval for each of the parameters
}
\description{
This function gives the Confidence interval (2.5% and 97.5%) for the fitted parameters
}

