% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.blm}
\alias{confint.blm}
\title{Confidence intervals of a Blm class}
\usage{
\method{confint}{blm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{Blm class object}

\item{parm}{A specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered. If the name of the response variable is given, then the confidence intervals of the fitted values values, or new predicted values if a data frame of new observations is given.}

\item{level}{Confidence interval level of interest, corresponds to a percentage.}

\item{...}{Additional parameters}
}
\value{
Matrix of confidence intervals, or a list of 2 vectors with upper-bound and lower-bound values for the fitted, or predicted variables.
}
\description{
Retrieve confidence intervals of the coefficients, or from observations
}

