% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{construct_model_matrix}
\alias{construct_model_matrix}
\title{Construct model matrix}
\usage{
construct_model_matrix(formula, Data = NULL)
}
\arguments{
\item{formula}{an object of type formula}

\item{Data}{an object of type dataframe}
}
\value{
a model matrix with data either from df or from the global environment
}
\description{
This functions constructs a model matrix with or without provided data
}

