% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{contains}
\alias{contains}
\title{Check if a list / object contains a value}
\usage{
contains(x, val)
}
\arguments{
\item{x}{object or list}

\item{val}{character value of the name of the object/list element to check}
}
\value{
TRUE if the value is present, FALSE if not.
}
\description{
Check if a list / object contains a value
}
\examples{
x <- list("a"=1, "b"=2)
contains(x, "a")

}
