% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ctmat}
\alias{ctmat}
\title{Cosine transform}
\usage{
ctmat(n)
}
\arguments{
\item{n}{Dimension of the matrix.}
}
\description{
Creates a cosine transform matrix of the given size.
}
\details{
The result F is a real orthonormal matrix \code{crossprod(F)=crossprod(t(F) = diag(n))} with first row a constant
and succesive rows \code{k} after that oscillating  with periods \code{2n/(k-1)}.  .
}

