% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{delete_posterior}
\alias{delete_posterior}
\title{Remove the posterior samples from a blm object.}
\usage{
delete_posterior(x)
}
\arguments{
\item{x}{blm object}
}
\value{
blm object minus posterior samples
}
\description{
Generally, there is no good reason to delete the posterior samples. The only situation in which this may occur if the user has already sampled the posterior and wants to change the sampling plan, which is locked after the posterior is sampled.
}
