% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.blm}
\alias{deviance.blm}
\title{Model Deviance for blm models}
\usage{
\method{deviance}{blm}(object, ...)
}
\arguments{
\item{object}{blm object the deviance}

\item{...}{Additional parameters.}
}
\value{
The deviance of the object
}
\description{
Returns the deviance of the blm model object
}

