% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{diagnostic_plots}
\alias{diagnostic_plots}
\title{Diagnostic Plots}
\usage{
diagnostic_plots(object, which = c(1:3), show_var = FALSE, id.n = 3,
  labels.id = names(residuals(object)), cex.id = 0.75, qqline = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{blm}} object.}

\item{which}{subset of the plots to display.}

\item{show_var}{display errorbars showing standard deviation on the 
residuals vs fitted plot.}

\item{id.n}{number of points to be labelled in each plot, starting with the
most extreme. Applies to plots 1 and 2.}

\item{labels.id}{vector of labels, from which the labels for extreme points
will be chosen. NULL uses observation numbers. Applies to plots 1 and 2.}

\item{cex.id}{magnification of point labels. Applies to plots 1 and 2.}

\item{qqline}{add qqline to the Q-Q plot.}

\item{...}{other arguments passed to the plot functions.}
}
\description{
Diagnostic plots for a \code{\link{blm}} object.
}
\details{
The following plots are available: 
 \enumerate{ 
   \item \code{\link{resid_vs_fitted_plot}} 
   \item \code{\link{scale_location_plot}} 
   \item \code{\link{qq_blm_plot}}
 }
}

