% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{directors}
\alias{directors}
\title{Compensation of 336 independent directors in S&P500 firms}
\format{A data frame containing 336 observations and 4 variables:
\describe{
   \item{Sector}{Sector the independent director works in}
   \item{Male}{Dummy indicating if the director is male}
   \item{Age}{Age of the director}
   \item{Compensation}{Compensation, in thousands GBR}
}}
\source{
Boardex, 2014
}
\usage{
directors
}
\description{
This dataset contains a cross-section of the compensation in GBR of 336 independent directors at S&P500 firms. The data were collected from Boardex (http://corp.boardex.com/) in 2014.
}
\keyword{datasets}
