% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution.blm}
\alias{distribution.blm}
\title{Distribution}
\usage{
\method{distribution}{blm}(x)
}
\arguments{
\item{x}{\code{blm} object.}
}
\description{
Distribution of a \code{blm} object.
}

