% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{distribution.default}
\alias{distribution.default}
\title{Distribution}
\usage{
\method{distribution}{default}(x)
}
\arguments{
\item{x}{object containing a distribution}
}
\description{
Default for generic method returning the distribution of an object.
}

