% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{draw_value}
\alias{draw_value}
\title{Draw a value from a prior distribution}
\usage{
draw_value(x)
}
\arguments{
\item{x}{prior object}
}
\value{
random variate drawn from a Normal or Gamma distribution
}
\description{
Draw a value from a prior distribution
}
