% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate_convergence_diagnostics}
\alias{evaluate_convergence_diagnostics}
\title{Print diagnostics}
\usage{
evaluate_convergence_diagnostics(x)
}
\arguments{
\item{x}{blm object}
}
\value{
prints summary of convergence diagnostics to R console
}
\description{
Print diagnostics
}
