% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate_effective_sample_size}
\alias{evaluate_effective_sample_size}
\title{Calculate the effective sample size}
\usage{
evaluate_effective_sample_size(x)
}
\arguments{
\item{x}{a blm object}
}
\value{
prints effective sample size information to screen
}
\description{
Calculate the effective sample size
}
