% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate_model_BF}
\alias{evaluate_model_BF}
\title{Evaluate the model Bayes Factor}
\usage{
evaluate_model_BF(x)
}
\value{
blm object with additional slot called 'model_BF' containing Bayes Factor information
}
\description{
The model Bayes Factor computes the evidence of the model versus the evidence found in the intercept-only model. To compute the model Bayes Factor, the user needs to explicitly state that they want to compute the null model (see examples). This Bayes Factor value is approximated using Wagemakers' approximation (see below for citation).
}
\examples{
# Load the directors data
data("directors")
# Fit a blm model
fit <- blm("Compensation ~ Age", data=directors) \%>\%
   # Compute the null model next to the desired model
   compute_null_model() \%>\%
   # Sample posterior data
   sample_posterior()
# Check the convergence of the null model
plot(fit, "nullmodel")
# The Bayes Factor is in the summary
summary(fit)

}
\seealso{
Wagenmakers, E. J. (2007). A practical solution to the pervasive problems of p values. Psychonomic bulletin & review, 14(5), 779-804.
}
