% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate_model_fit}
\alias{evaluate_model_fit}
\title{Assess blm model fit}
\usage{
evaluate_model_fit(x)
}
\arguments{
\item{x}{blm object}
}
\value{
prints summary of model fit (DIC) to the R console
}
\description{
Assess blm model fit
}
