% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate_outliers}
\alias{evaluate_outliers}
\title{Evaluate possible outliers}
\usage{
evaluate_outliers(x)
}
\arguments{
\item{x}{a blm object}
}
\value{
object of class "outliers" is added to the blm object
}
\description{
Evaluate possible outliers
}
\seealso{
Lynch, S. M. (2007). Introduction to applied Bayesian statistics and estimation for social scientists. Springer Science & Business Media. pp.178-182
}
