% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate_ppc}
\alias{evaluate_ppc}
\title{Set up posterior predictive checks}
\usage{
evaluate_ppc(x, ...)
}
\arguments{
\item{x}{blm object}
}
\value{
prints summary of the ppc to the R console
}
\description{
Set up posterior predictive checks
}
