% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posteriors.R
\name{fit_model}
\alias{fit_model}
\title{Fit the model}
\usage{
fit_model(model, data, prior, beta)
}
\arguments{
\item{model}{the model to fit}

\item{data}{the data with response}

\item{prior}{the prior distribution of model}

\item{beta}{the beta value}
}
\value{
a list with means for each parameter and a covariance matrix
}
\description{
This function fits the model to the data given the prior distribution and the beta.
}

