% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.blm}
\alias{fitted.blm}
\title{Extracted fitted response values}
\usage{
\method{fitted}{blm}(object, ...)
}
\arguments{
\item{object}{blm model object}

\item{...}{Additional pararmeters, forwarded to predict.blm()}
}
\value{
vector of the fitted response variables
}
\description{
Returns the fitted response values from the model, given the mean values they would correspond to
}

