% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted}
\alias{fitted}
\title{Bayesian linear model.}
\usage{
fitted(x, ...)
}
\arguments{
\item{x}{A formula describing the model.}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\value{
A fitted model.
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}

