% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_parameter_names}
\alias{get_parameter_names}
\title{Retrieve mapping from variable to parameter names}
\usage{
get_parameter_names(x)
}
\arguments{
\item{x}{a blm object}
}
\value{
prints mapping from parameter to variable names to the console
}
\description{
blm objects will mostly emit parameter names (b0, b1, etc) instead of variable names. This convenience function returns the mapping from parameter names to variable names.
}
