% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_shared.R
\name{get_value}
\alias{get_value}
\alias{get_value.blm}
\alias{get_value.DIC}
\alias{get_value.R2}
\alias{get_value.priors}
\alias{get_value.posterior}
\alias{get_value.chain}
\alias{get_value.ppc}
\alias{get_value.sampler}
\title{Get a value from an S3 object}
\usage{
get_value(x, var)

\method{get_value}{blm}(x, var)

\method{get_value}{DIC}(x, var)

\method{get_value}{R2}(x, var)

\method{get_value}{priors}(x, var)

\method{get_value}{posterior}(x, var)

\method{get_value}{chain}(x, var)

\method{get_value}{ppc}(x, var)

\method{get_value}{sampler}(x, var)
}
\arguments{
\item{x}{object with a get_value() method}

\item{var}{string. name of the value to retrieve from the object.}
}
\description{
Get a value from an S3 object
}
