\name{grad}
\docType{data}
\alias{grad}

\title{Data set on admission to graduate school}

\description{Admission outcomes by academic performance and univeristy characteristics}
\usage{
grad
}
\format{
  \tabular{ll}{
    admit:\tab indicator for admission to grad school \cr
    gre:\tab GRE score \cr
    gpa: \tab grade point average \cr
    rank: \tab prestige of undergraduate institution (1 highest, 4 lowest)\cr
  }
}
\source{
\url{http://www.ats.ucla.edu/stat/r/dae/logit.htm}
}
\keyword{datasets}
