% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{homoskedast_check}
\alias{homoskedast_check}
\title{Check a posterior sample for homoskedasticity assumption}
\usage{
homoskedast_check(x)
}
\arguments{
\item{x}{ppc object}
}
\value{
posterior predictive check for homoskedasticity assumption
}
\description{
Check a posterior sample for homoskedasticity assumption
}
