% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{hypothesis}
\alias{hypothesis}
\title{Pass a hypothesis and parse it. THen return object of class 'hypothesis'}
\usage{
hypothesis(hypothesis_user, parameters)
}
\arguments{
\item{hypothesis_user}{a hypothesis passed by the user.}

\item{parameters}{parameters used in the model (e.g. b0, b1, b2 ...)}
}
\value{
object of class 'hypothesis' containing the original hypothesis and its parsed and checked version
}
\description{
Hypotheses are used to do Bayesian hypothesis testing using Bayes' Factors
}
