% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{independence_check}
\alias{independence_check}
\title{Check a posterior sample for independence of errors assumption}
\usage{
independence_check(x, ...)
}
\arguments{
\item{x}{ppc object}
}
\value{
posterior predictive check for independence of errors assumption
}
\description{
Check a posterior sample for independence of errors assumption
}
