% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_plots.R
\name{kernel_density}
\alias{kernel_density}
\title{Density Plot of Multivariate Normal Distribution}
\usage{
kernel_density(object, n = 10000, ...)
}
\arguments{
\item{object}{a \code{mvnd} object}

\item{n}{sampling depth used for kernel density estimation.}

\item{...}{additional arguments passed to the plot function.}
}
\description{
Plots a kernel density for a \code{mvnd} object.
}
\details{
Draws random samples from a multivariate normal distribution of
  \code{\link{mvnd}} object and plots the kernel density of those.
  Requires the mvrnorm from the MASS package for random sampling. If present,
  the package LSD is used for kernel density coloring. ... argments are
  passed to generic plot or \code{\link[LSD]{heatscatter}} functions if LSD
  package is installed.
}
\examples{
x <- rnorm(100)
b <- 1.3
w0 <- 0.2 ; w1 <- 3

y <- rnorm(100, mean = w0 + w1 * x, sd = sqrt(1/b))
model <- blm(y ~ x, prior = NULL, beta = b, 
               data = data.frame(x=x, y=y)) 

\dontrun{
par(mfrow=c(1,2))
kernel_density(model$prior, xlim=c(-4,4), ylim=c(-4,4), main='prior')    
kernel_density(model$posterior, xlim=c(-4,4), ylim=c(-4,4), main='posterior')
par(mfrow=c(1,1)) 
} 
               
}

