% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lagts}
\alias{lagts}
\title{Create regression rhs matrix with lags}
\usage{
lagts(xts, lags)
}
\arguments{
\item{xts}{The series we will be using.  An \code{mts} object.}

\item{lags}{Either a list of integer vectors giving the lags of each variable to appear in the output,
or a single integer vector, which will be re-used as the list of lags for all variables.}
}
\description{
The general case is that xts is an mts object and lags a list of length equal to the
number of series in \code{xts}.  The lags specified in lags are added to the xts object as
additional columns, and the resulting mts object is trimmed to eliminate NA's that would
otherwise arise. The series in the returne object retain their original names, with integers
appended indicating how far they have been lagged.
In experimenting with regression models with varying lag lengths, it is a good idea to set up
a single mts object with this program at the start, so that the same sample size is being
used in comparing models with different lags.
Note that the series are trimmed to eliminate all NA's, so that an NA in the middle of any
series in xts  will result in extreme truncation.
xts can also be a single time series.  lags can be just a single vector of lags, in which case
all series in xts are included with the same lags  Though this function allows non-contiguous
lags, \code{\link{tsregPrior}} allows only contiguous lags..
}

