% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_misc.R
\name{LaTeX_summary}
\alias{LaTeX_summary}
\title{Create a LaTeX table for blm model summaries}
\usage{
LaTeX_summary(...)
}
\arguments{
\item{...}{one or more blm models}
}
\description{
Create a LaTeX table for blm model summaries
}
\examples{
LaTeX_summary(bfit1, bfit2, bfit3)
}
\seealso{
this function is based on output created by the stargazer library
}
