% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posteriors.R
\name{make_model_matrix}
\alias{make_model_matrix}
\title{Make a model matrix}
\usage{
make_model_matrix(model, data)
}
\arguments{
\item{model}{the model the data will be fitted to}

\item{data}{the data}
}
\value{
a model matrix
}
\description{
This function make a model matrix for the data
}

