% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior.R
\name{make_prior}
\alias{make_prior}
\title{Make the prior distribution}
\usage{
make_prior(model, alpha)
}
\arguments{
\item{model}{A formula describing the model.}

\item{alpha}{Hyper-parameter used in the the prior}
}
\value{
a prior distribution, list of mu and Sigma
}
\description{
Makes a prior distribution, according to the corresponding alpha parameter
}

