% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posteriors.R
\name{make_responseless_matrix}
\alias{make_responseless_matrix}
\title{Make a responseless model matrix}
\usage{
make_responseless_matrix(model, data)
}
\arguments{
\item{model}{the model the data will be fitted to}

\item{data}{the data without a response variable}
}
\value{
a model matrix
}
\description{
This function make a responseless model matrix for the data.
}

