% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{model.matrix.blm}
\alias{model.matrix.blm}
\title{Model Matrix}
\usage{
\method{model.matrix}{blm}(object, ...)
}
\arguments{
\item{object}{a \code{blm} object.}

\item{...}{other arguments (currently ignored).}
}
\description{
Model matrix for \code{\link{blm}} object.
}

