% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model_fit}
\alias{model_fit}
\title{Assess blm model fit}
\usage{
model_fit(x)
}
\arguments{
\item{x}{blm object}
}
\value{
prints summary of model fit (DIC) to the R console
}
\description{
Assess blm model fit
}
