% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix_responseless.R
\name{model_matrix_responseless}
\alias{model_matrix_responseless}
\title{Model matrix responseless.}
\usage{
model_matrix_responseless(model, ...)
}
\arguments{
\item{model}{A formula describing the model.}

\item{...}{Additional data, for example a data frame.}
}
\value{
Model matrix
}
\description{
Fits a model with a responseless matrix, given as a formula, optionally with data provided through the "..." parameter.
}

