% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{model_rmse}
\alias{model_rmse}
\title{Check a posterior sample for RMSE assumption}
\usage{
model_rmse(x)
}
\arguments{
\item{x}{ppc object}
}
\value{
posterior predictive check for RMSE assumption
}
\description{
Check a posterior sample for RMSE assumption
}
