% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\docType{package}
\name{ms_blm}
\alias{ms_blm}
\alias{ms_blm-package}
\title{blm: A package for bayesian linear regression.}
\description{
The \code{blm} package is for construction and analysis of bayesian linear
regression models.
}
\section{blm functions}{

blm
update.blm
print.blm
summary.blm
fitted.blm
predict.blm
summary.blm
resid.blm
coef.blm
deviance.blm
plot.blm
}

