% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mv_dist.means}
\alias{mv_dist.means}
\title{Means}
\usage{
mv_dist.means(d, ...)
}
\arguments{
\item{d}{object of class \code{\link{mv_dist}}.}

\item{...}{other arguments (currently ignored).}
}
\description{
Means of multivariate normal distribution object of class \code{mv_dist}.
}

