% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mv_dist}
\alias{mv_dist}
\title{Multivariate distribution}
\usage{
mv_dist(means, covar, var_names, ...)
}
\arguments{
\item{means}{vector of means of the variables}

\item{covar}{covariance matrix of the variables}

\item{var_names}{optional names of the variables}

\item{...}{other arguments (currently ignored).}
}
\value{
An object of class \code{distribution} containing: 
\describe{
   \item{means}{vector of means}
   \item{covar}{covariance matrix}
}
}
\description{
Means and covariance matrix of a multivariate normal distribution .
}
\examples{
mv_dist(c(0,1),matrix(c(2,3,3,2),ncol=2))

}

