% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mv_dist.set_var_names}
\alias{mv_dist.set_var_names}
\title{Set variable Names from a Multivariate Normal Distribution}
\usage{
mv_dist.set_var_names(d, var_names, ...)
}
\arguments{
\item{d}{object of class \code{\link{mv_dist}}.}

\item{var_names}{names of the variables of the distributions.}

\item{...}{other arguments (currently ignored).}
}
\description{
Set the names of the variables of a multivariate normal distribution object
of class \code{mv_dist}.
}

