% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mv_dist.subset}
\alias{mv_dist.subset}
\title{Subset a Multivariate Normal Distribution}
\usage{
mv_dist.subset(object, subset)
}
\arguments{
\item{object}{an object of class \code{mv_dist}.}

\item{subset}{names or indices of variables to extract from mv_dist.}
}
\description{
Return subset of variables of a multivariate normal distribution object of
class \code{mv_dist}.
}
\examples{
d <- mv_dist(c(0,1),matrix(c(2,3,3,2),ncol=2), c('a', 'b'))
mv_dist.subset(d, 1)
mv_dist.subset(d, 'b')

}

