% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mv_dist.var_names}
\alias{mv_dist.var_names}
\title{Variable Names from a Multivariate Normal Distribution}
\usage{
mv_dist.var_names(d, ...)
}
\arguments{
\item{d}{object of class \code{\link{mv_dist}}.}

\item{...}{other arguments (currently ignored).}
}
\description{
Names of the variables of a multivariate normal distribution object of class 
 \code{mv_dist}.
}

