% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mvnd.covar}
\alias{mvnd.covar}
\title{Covariance}
\usage{
mvnd.covar(d, ...)
}
\arguments{
\item{d}{object of class \code{\link{mvnd}}.}

\item{...}{other arguments (currently ignored).}
}
\description{
Covariance matrix of multivariate normal distribution object of class 
 \code{mvnd}.
}

