% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mvnd.means}
\alias{mvnd.means}
\title{Means}
\usage{
mvnd.means(d, ...)
}
\arguments{
\item{d}{object of class \code{\link{mvnd}}.}

\item{...}{other arguments (currently ignored).}
}
\description{
Means of multivariate normal distribution object of class \code{mvnd}.
}

