% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mvnd.subset}
\alias{mvnd.subset}
\title{Subset a Multivariate Normal Distribution}
\usage{
mvnd.subset(object, subset)
}
\arguments{
\item{object}{an object of class \code{mvnd}.}

\item{subset}{names or indices of variables to extract from mvnd.}
}
\description{
Return subset of variables of a multivariate normal distribution object of
class \code{mvnd}.
}
\examples{
d <- mvnd(c(0,1),matrix(c(2,3,3,2),ncol=2), c('a', 'b'))
mvnd.subset(d, 1)
mvnd.subset(d, 'b')

}

