% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mvnd.var_names}
\alias{mvnd.var_names}
\title{Variable Names from a Multivariate Normal Distribution}
\usage{
mvnd.var_names(d, ...)
}
\arguments{
\item{d}{object of class \code{\link{mvnd}}.}

\item{...}{other arguments (currently ignored).}
}
\description{
Names of the variables of a multivariate normal distribution object of class 
 \code{mvnd}.
}

