% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{norm_of_errors_check}
\alias{norm_of_errors_check}
\title{Check a posterior sample for the assumption that errors are normally distributed with mean mu and sd sigma}
\usage{
norm_of_errors_check(x, ...)
}
\arguments{
\item{x}{ppc object}
}
\value{
posterior predictive check on normal distribution of errors assumption
}
\description{
Check a posterior sample for the assumption that errors are normally distributed with mean mu and sd sigma
}
