% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{normality_check}
\alias{normality_check}
\title{Check a posterior sample for normality assumption}
\usage{
normality_check(x, ...)
}
\arguments{
\item{x}{ppc object}
}
\value{
posterior predictive check for normality assumption
}
\description{
Check a posterior sample for normality assumption
}
