% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_plots.R
\name{parameter_kernel}
\alias{parameter_kernel}
\title{Density Plot of Parameters}
\usage{
parameter_kernel(object, n = 10000, sample_from_prior = FALSE, ...)
}
\arguments{
\item{object}{a \code{blm} object}

\item{n}{sampling depth used for kernel density estimation.}

\item{sample_from_prior}{draw samples from the prior distribution?}

\item{...}{additional arguments passed to the plot function.}
}
\description{
Plots a kernel density for the distribution of the parameters of a \code{blm} object.
}
\details{
Draws random samples from the posterior (or prior if sample_from_prior = TRUE) distribution of \code{\link{blm}} object and plots the kernel density of those. Requires the mvrnorm from the MASS package for random sampling. If present, the package LSD is used for kernel density coloring. ... argments are passed to generic plot or \code{\link[LSD]{heatscatter}} functions if LSD package is installed.
}
\examples{
x <- rnorm(100)
b <- 1.3
w0 <- 0.2 ; w1 <- 3

y <- rnorm(100, mean = w0 + w1 * x, sd = sqrt(1/b))
model <- blm(y ~ x, prior = NULL, beta = b, 
               data = data.frame(x=x, y=y)) 

\dontrun{
par(mfrow=c(1,2))
parameter_kernel(model, sample_from_prior = TRUE, xlim=c(-4,4), ylim=c(-4,4))    
parameter_kernel(model, sample_from_prior = FALSE, xlim=c(-4,4), ylim=c(-4,4))
par(mfrow=c(1,1)) 
} 
               
}

