% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_analysis.R
\name{pblm}
\alias{pblm}
\title{Probability of Response}
\usage{
pblm(object, a_prior = 1, b_prior = 1, use_log = TRUE)
}
\arguments{
\item{object}{a \code{\link{blm}} object.}

\item{a_prior}{prior value for parameter a of inv. gamma distribution.}

\item{b_prior}{prior value for parameter b of inv. gamma distribution.}

\item{use_log}{compute the log-probability.}
}
\description{
Probability of observing the response given a (bayesian) linear regression
model.
}
\details{
Computes as described in 
  \url{https://en.wikipedia.org/wiki/Bayesian_linear_regression}. By default
  the log-probability is computed as non log-transformed probabilites are not
  solvable.
}

