% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot blm object}
\usage{
\method{plot}{blm}(x, ...)
}
\arguments{
\item{x}{blm object}

\item{...}{Additional parameters like ggplot}
}
\value{
a figure
}
\description{
Plots an blm object
}

