% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ppc.R
\name{plot.ppc}
\alias{plot.ppc}
\title{Plot posterior predictive samples}
\usage{
\method{plot}{ppc}(ppc, type = c("normality", "heteroskedasticity",
  "independence"))
}
\arguments{
\item{ppc}{a posterior predictive check object (ppc)}

\item{type}{one of 'normality', 'heteroskedasticity' or 'independence'}
}
\description{
Plotting is only possible if you choose to return 'return_samples=TRUE' in \code{evaluate_ppc()}
}
\examples{
data("directors")
fit <- blm("Compensation ~ Age", data=directors) \%>\%
   sample_posterior()
# Calculate PPC
fit <- fit \%>\%
   evaluate_ppc(return_samples=TRUE)
# Plot
plot(fit \%>\% get_value('ppc'), 'normality')

}
