% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plots}
\alias{plots}
\title{Plotting the variables}
\usage{
plots(object, ...)
}
\arguments{
\item{object}{A fitted model obtained from blm function}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\description{
This function plots the response variable and each of the predictor variable one by one.
}

