% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{posterior}
\alias{posterior}
\title{Sets up posterior class}
\usage{
posterior(samples)
}
\value{
list containing matrices of chain values (dims: iterations x coef + 1) of object posterior
}
\description{
Sets up posterior class
}
