% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{posterior_predictive_checks}
\alias{posterior_predictive_checks}
\title{Set up posterior predictive checks}
\usage{
posterior_predictive_checks(x, ...)
}
\arguments{
\item{x}{blm object}
}
\value{
prints summary of the ppc to the R console
}
\description{
Set up posterior predictive checks
}
