% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{precision}
\alias{precision}
\title{Precision}
\usage{
precision(object, ...)
}
\arguments{
\item{object}{a \code{blm} object.}

\item{...}{other arguments (currently ignored).}
}
\value{
The precision used for the \code{\link{blm}} object.
}
\description{
Precision of error of a bayesian model.
}
\examples{
x <- rnorm(10)
b <- 1.3
w0 <- 0.2 ; w1 <- 3
y <- rnorm(10, mean = w0 + w1 * x, sd = sqrt(1/b))

#precision provided as argument
model <- blm(y ~ x, beta = b)
precision(model)

#precision estimated from the data
model2 <- blm(y ~ x)
precision(model2)

}

