% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.blm}
\alias{predict.blm}
\title{Predict new data from blm model}
\usage{
\method{predict}{blm}(object, variances = FALSE, ...)
}
\arguments{
\item{object}{Blm object}

\item{variances}{Whether to return variances aswell}

\item{...}{Additional parameters, in particular new data to predict from.}
}
\value{
predicted values
}
\description{
Predict new data points from a blm model, if no new data is given returns the fitted response variables. If new data is given, it gives predicted response variables.
}

