% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Posteriors.R
\name{predict_response}
\alias{predict_response}
\title{Predicting responses}
\usage{
predict_response(new_data, fitted, model, beta)
}
\arguments{
\item{new_data}{the new values for the predictor variables}

\item{fitted}{the fitted model, as outputted from fit_model}

\item{model}{the model from the fit}

\item{beta}{the beta value}
}
\value{
a list with a vector of means (predicted responses) and a vector of corresponding variances
}
\description{
This function predict responses given new predictor values and the mean and covariance for the fitted model
}

