% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_analysis.R
\name{print.anova.blm}
\alias{print.anova.blm}
\title{Print}
\usage{
\method{print}{anova.blm}(x, ...)
}
\arguments{
\item{x}{a \code{\link{anova.blm}} object.}

\item{...}{other arguments (currently ignored).}
}
\description{
Print a anova.blm object
}

