% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.blm}
\alias{print.blm}
\title{Print}
\usage{
\method{print}{blm}(x, ...)
}
\arguments{
\item{x}{The Bayesian linear regression model}

\item{...}{Additional data, for example a data frame.}
}
\value{
print
}
\description{
Prints the fitted model and its coefficients
}

