\name{print}
\docType{methods}
\alias{print}
\alias{print,blm-method}
\alias{print,lexpit-method}

\title{Print coefficients of \code{blm} and \code{lexpit} model fit.}

\description{

Prints the regression coefficients of the fit of a \code{blm} or \code{lexpit} model.

}

\section{Methods}{
\describe{

\item{print}{\code{signature(x = "blm")}:
Call and coefficient estimates.
  }

\item{print}{\code{signature(x = "lexpit")}:
Call and coefficient estimates.
  }
 }
}

\author{Stephanie Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
