% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.R
\name{print.summary.blm}
\alias{print.summary.blm}
\title{Print summary.blm}
\usage{
\method{print}{summary.blm}(x)
}
\arguments{
\item{x}{a summary.blm object}
}
\value{
nothing
}
\description{
Prints the summary details of summary.blm object
}

