% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{prints}
\alias{prints}
\title{Print}
\usage{
prints(object)
}
\arguments{
\item{object}{A fit model obtained from blm function}
}
\description{
This function prints how the fitting function was called and what the fitted coefficients are.
}

