% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{prior}
\alias{prior}
\title{Sets up a prior class}
\usage{
prior(density, ...)
}
\arguments{
\item{density}{either Normal or Gamma density}
}
\value{
prior object
}
\description{
Every blm object contains an S3 class called 'priors'. This object is used to store information about priors and has its own methods (mainly used for internal purposes). The user can specify priors by means of the function 'set_priors()'
}
