% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{qq_blm_plot}
\alias{qq_blm_plot}
\title{Q-Q Plot}
\usage{
qq_blm_plot(object, qqline = TRUE, ...)
}
\arguments{
\item{object}{a \code{\link{blm}} object.}

\item{qqline}{add qqline to the plot.}

\item{...}{other arguments passed to qqplot.}
}
\description{
Q-Q Plot for a \code{\link{blm}} object.
}
\details{
Q-Q plots the sqrt(| residuals |) against fitted values. This
  transformation of the residuals is supposed to diminish skewness (see
  \code{\link{plot.lm}}. Note: In contrast to the \code{\link{plot.lm}}
  implementation, the residuals here are taken for sqrt(| residuals |) and
  not the standardized residuals.
}

