% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_plots.R
\name{resid_vs_fitted_plot}
\alias{resid_vs_fitted_plot}
\title{Plot of Residuals against Fitted Values}
\usage{
resid_vs_fitted_plot(object, show_var = FALSE, id.n = 3,
  labels.id = names(residuals(object)), cex.id = 0.75, ...)
}
\arguments{
\item{object}{a \code{\link{blm}} object.}

\item{show_var}{display errorbars showing standard deviation on the 
residuals.}

\item{id.n}{number of points to be labelled in each plot, starting with the
most extreme.}

\item{labels.id}{vector of labels, from which the labels for extreme points
will be chosen. NULL uses observation numbers.}

\item{cex.id}{magnification of point labels.}

\item{...}{other arguments passed to plot.}
}
\description{
Plots residuals against fitted values for a \code{\link{blm}} object.
}
\details{
Standard deviation of the residuals is same as for the fitted
  values.
}

