% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.blm}
\alias{residuals.blm}
\title{Residuals from the blm object}
\usage{
\method{residuals}{blm}(object, ...)
}
\arguments{
\item{object}{Blm object}

\item{...}{Additional parameters}
}
\value{
a vector of the residuals
}
\description{
This function returns the residuals of the fit. That is the difference between predicted values and observed values for the response variable.
}

