% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{residuals}
\alias{residuals}
\title{Bayesian linear model.}
\usage{
residuals(x, ...)
}
\arguments{
\item{x}{A formula describing the model.}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\value{
A fitted model.
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}

