% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runRSN.R
\name{runRSN}
\alias{runRSN}
\title{Robust Spline Normalisation}
\usage{
runRSN(eset, refSet, minValue = 0, ...)
}
\arguments{
\item{eset}{an ExpressionSet object}

\item{refSet}{an ExpressionSet object}

\item{minValue}{TODO}

\item{...}{TODO
The function can be run multicore by running if registered via the doMC or doParallel package}
}
\description{
A wrapper around the lumi::rsn() function
}
\examples{
set.seed(1)
#TODO!

}
