% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Priors.R
\name{sample_from_prior}
\alias{sample_from_prior}
\title{Draw a sample from a prior distribution for a linear model}
\usage{
sample_from_prior(n, alpha)
}
\arguments{
\item{n}{the wanted number of samples}

\item{alpha}{the alpha value}
}
\value{
a matrix with the estimated weights w0 and w1 as columns
}
\description{
This function make sample from the prior distribution
}

