% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample the posterior distribution}
\usage{
sample_posterior(x)
}
\arguments{
\item{x}{blm object}
}
\value{
blm object containing sampled posterior
}
\description{
Calling the \code{sample_posterior()} function executes a blm sampling plan. The user should set priors, initial values etc. before executing this function, because by executing it the user will 'lock' the sampling plan. That is, no changes may be made except for changing the 'burn' parameter.
}
\details{
For details on the implementation of the Gibbs sampler, see https://tinyurl.com/y5vk9x35. For details on the implementation of the Metropolis-Hastings sampler, see https://tinyurl.com/y3zp3l3w.
}
\examples{
data("directors")
fit <- blm("Compensation ~ Age", data=directors) \%>\%
   sample_posterior()

}
