% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{scale_location_plot}
\alias{scale_location_plot}
\title{Scale-Location Plot}
\usage{
scale_location_plot(object, ...)
}
\arguments{
\item{object}{a \code{\link{blm}} object.}

\item{...}{other arguments passed to plot.}
}
\description{
Scale-Location Plot for a \code{\link{blm}} object.
}
\details{
Scale-location plots the sqrt(| residuals |) against fitted values.
  This is supposed to diminish skewness. Note: In contrast to the
  \code{\link{plot.lm}} implementation, the residuals here are taken for
  sqrt(| residuals |) and not the standardized residuals.
}

