% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_initial_values}
\alias{set_initial_values}
\title{Set initial values for a chain}
\usage{
set_initial_values(x, ...)
}
\arguments{
\item{x}{blm object}

\item{...}{named argument containing (1) a chain name (e.g. 'chain_1') and (2) a list of initial values. See examples.}
}
\value{
updated blm object
}
\description{
Set initial values for a chain
}
\examples{
bfit <- set_initial_values(bfit, chain_1 = list("b" = c(2,5,3,4,5), sigma=0.1))

}
