% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sampler.R
\name{set_options.sampler}
\alias{set_options.sampler}
\title{Updating sampling options}
\usage{
\method{set_options}{sampler}(x, chains = 1, iterations = 10000,
  burn = 1000, thinning = 1, varnames, priors)
}
\description{
Updating sampling options
}
