% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_prior}
\alias{set_prior}
\title{Change priors for available parameters}
\usage{
set_prior(x, par, ...)
}
\arguments{
\item{par}{name of the coefficient, given as b0 (intercept) or b1, ... for coefficients.}

\item{...}{other arguments passed to function. In the case of intercept and coefficients, these are 'mu' and 'sd' (prior means and variances). In the case of the residual variance, the rate and shape/scale parameter must be passed as 'alpha' and 'beta'}
}
\value{
updated blm object with new priors
}
\description{
Change priors for available parameters
}
\examples{
bfit <- set_prior(bfit,"b0", mu=40, sd=7)
bfit <- set_prior(bfit, "sigma", alpha=2, beta=1)
}
