% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_priors}
\alias{set_priors}
\title{Set priors for a blm object}
\usage{
set_priors(x, ...)
}
\arguments{
\item{x}{blm object}

\item{...}{optional arguments needed to specify a prior}
}
\value{
blm object with new priors
}
\description{
Set priors for a blm object
}
