% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_sampler.R
\name{set_sampler.sampler}
\alias{set_sampler.sampler}
\title{Updating the type of sampler (e.g. gibbs or mh)}
\usage{
\method{set_sampler}{sampler}(x, label, type = c("Gibbs", "MH"),
  zeta = 0.25)
}
\description{
Updating the type of sampler (e.g. gibbs or mh)
}
