% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_sampling_options}
\alias{set_sampling_options}
\title{Sample the posterior distribution}
\usage{
set_sampling_options(x, ...)
}
\arguments{
\item{blm}{blm object}

\item{chains}{number of chains to run}

\item{iterations}{number of iterations per chain}

\item{burn}{number of burn-in iterations}

\item{julia}{use julia for computations?}
}
\value{
updated blm object
}
\description{
Sample the posterior distribution
}
