% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{set_value}
\alias{set_value}
\title{Set a value for an S3 object}
\usage{
set_value(x, var, val)
}
\arguments{
\item{x}{object with a set_value() method}

\item{var}{string. name of the value you want to change}

\item{val}{new value for var}
}
\description{
Set a value for an S3 object
}
