% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{subset.mv_dist}
\alias{subset.mv_dist}
\title{Subset a Multivariate Normal Distribution}
\usage{
\method{subset}{mv_dist}(object, subset)
}
\arguments{
\item{object}{an object of class \code{mv_dist}}

\item{subset}{names or indices of variables to extract from mv_dist}
}
\description{
Return ubset of variables of a multivariate normal distribution object of 
 class \code{mv_dist}.
}
\examples{
d <- mv_dist(c(0,1),matrix(c(2,3,3,2),ncol=2), c('a', 'b'))
subset.mv_dist(d, 1)
subset.mv_dist(d, 'b')

}

