% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.blm}
\alias{summary.blm}
\title{Summary of blm object}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{blm object}

\item{...}{Additional variables}
}
\value{
summary.blm class, containing the summary statistics
}
\description{
Creates summary statistics of the blm model given.
}

