% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_misc.R
\name{summary.DIC}
\alias{summary.DIC}
\title{Summary method for DIC}
\usage{
\method{summary}{DIC}(x, ...)
}
\arguments{
\item{x}{DIC object from blm library}

\item{...}{DIC of the intercept-only model (if available)}
}
\description{
Summary method for DIC
}
