% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_theme.R
\name{theme_blm}
\alias{theme_blm}
\title{Custom theme for blm plots}
\usage{
theme_blm(text_size = 12, line_size = text_size/170,
  rect_size = text_size/170)
}
\arguments{
\item{text_size}{text size for blm plots}

\item{line_size}{size of the axis lines}

\item{rect_size}{sizs of the background lines}
}
\value{
adds blm theme to a plot generated by ggplot2
}
\description{
Custom theme for blm plots
}
\seealso{
https://www.r-bloggers.com/custom-themes-in-ggplot2/
}
