% Generated by roxygen2 (4.0.2): do not edit by hand
\name{trimts}
\alias{trimts}
\title{Trim a multiple time series object to eliminate NA's}
\usage{
trimts(y)
}
\arguments{
\item{y}{The mts object to be trimmed.}
}
\description{
Using \code{cbind} to combine lagged values of time series
creates NA's at the start and end.  This can be avoided by
using \code{ts.intersect} instead of cbind.  But if one starts
with an mts object with series of different time spans, the
NA's will be there, and ts.intersect ignores them.
}

