% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tsregPrior}
\alias{tsregPrior}
\title{Priors for time series regression}
\usage{
tsregPrior(vlist, dvname = "y", lags = rep(0, length(vlist)), ldv = 1,
  scale, bbar = NULL, smooth, damp, vmeans, erratio)
}
\arguments{
\item{vlist}{Character vector of names of the rhs variables}

\item{dvname}{Name of dependent variable}

\item{lags}{How many lags of each variable.  Can be a single integer,
        if all variables appear with the same number of lags, or a
        vector of integers, one for each variable. Note that this
        program does not care whether, when \code{lags[iv]=3}, this means that
        lags 1 to 3, 0 to 2, or 10 to 12 are included.}

\item{scale}{A vector, of the length of \code{vlist} plus 1, of
         reasonable guesses as to the standard deviations of
         the variables.  The last element scales the levels dummy that relates the constant to the
         other parameters, which should be 1.0, or smaller if guesses of variable means in vmeans are
         quite uncertain.}

\item{bbar}{mean of the regression coefficient vector.  Default of 1
         followed by zeros is reasonable when the first variable is a lagged dependent variable and this
         is a regression with a persistent dependent variable.  Last element
         is prior mean of constant (usually 0).}

\item{smooth}{The rate at which the scale of dummy observations drops as we go from low to
         high frequencies.  Should usually be in (0,1).  If close to zero, prior on the
         variable's effect is stronger at low than at high frequencies.  With smooth=1
         and damp=1, the prior just shrinks all coefficients toward their prior means
         symmetrically and independently.}

\item{damp}{The rate at which the dummy observations increase with increasing lag length. Should
         equal or exceed one if distant lags are more likely to be large than near-in lags.}

\item{vmeans}{A priori means for the variables.  They are used in forming the last dummy observation,
         connecting the constant to the other coefficients.  Could be set as sample means of initial conditions.
         Using full sample means is problematic: the contamination of prior by data may not be a big
         problem for stationary data, but could be substantially distorting for non-stationary data.}

\item{erratio}{ratio of error variance from parameter uncertainty to error variance from the
         residual for the first sample observation. Keeping this constant across models of
         different size avoids a prior that implies big models should fit much better.}
}
\value{
A list with components:
  \itemize{
    \item y. Dependent variable value for dummy observations.
    \item X. Dummy observations implementing the prior.
    \item scalefac. Amount by which the prior has been scaled to match \code{erratio}.
  }
}
\description{
Creates dummy observations for a prior symmetric in variables,
    favoring persistence.
}
\details{
The regression right-hand-side is assumed to contain lagged values, possibly
    of both dependent and independent variables.  By default it centers on a
    random walk with no effects of independent variables. The results can be used
    directly in \code{\link{blm}}

Note that the last column of \code{X} is the
           "constant", which will *not* be a column of ones.  Generating the sample
           \code{X} matrix using \code{lagts()} will not by itself create a constant term in the last column.
           \code{lagts()} does put the variables and lags in the same order as this function.  But you have
           to tack on the constant vector with \code{cbind()}.  \code{X} is a matrix with column names
           constructed from \code{vnames} (repeated for lags).  Note that \code{lagts()} allows non-sequential lists
           of lags.  This program allows different lag lengths, but always \code{1:lag[iv]}.
}
\seealso{
\code{\link{lagts}} for preparation of a data matrix in a form that is compatible with this
          function.  \code{\link{blm}} to combine the prior with data to obtain estimates.
}

