% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{update.blm}
\alias{update.blm}
\title{Update a blm}
\usage{
\method{update}{blm}(object, Data = NULL)
}
\arguments{
\item{object}{an object of class blm}

\item{Data}{an object of type dataframe}
}
\value{
an object of class blm
}
\description{
A function for refitting data with a new prior.
}

