% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update}
\usage{
update(model, prior, beta, ...)
}
\arguments{
\item{model}{A formula describing the model.}

\item{prior}{The prior distribution}

\item{beta}{A number}

\item{...}{Additional data, for example a data frame.}
}
\value{
Update function
}
\description{
Updates the prior distribution for inputing into the blm model
}

