% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_blm.R
\name{update_posterior}
\alias{update_posterior}
\alias{update_posterior.blm}
\title{Draw more samples from a sampled posterior distribution.}
\usage{
update_posterior(x, ...)

\method{update_posterior}{blm}(x, iterations = 1000)
}
\arguments{
\item{x}{blm object}

\item{...}{other options passed to function}

\item{iterations}{number of additional samples to draw from the posterior.}
}
\value{
blm object with updated count of posterior samples.
}
\description{
Draw more samples from a sampled posterior distribution.
}
\examples{
bfit <- update_posterior(blm, iterations=5000)

}
