% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_prior}
\alias{update_prior}
\title{Updating the Prior distribution into a Posterior distribution}
\usage{
update_prior(model, prior, beta, ...)
}
\arguments{
\item{model}{A formula describing the model}

\item{prior}{The prior distribution to be updated}

\item{beta}{Hyper-parameter beta, for uncertainty of the measurements.}

\item{...}{Additional data, for example a data frame.}
}
\value{
Posterior distribution, list of mu and Sigma
}
\description{
Updates the given prior distribution, using the observed data, to produce the posterior distribution
}

