\name{vcov}
\docType{methods}
\alias{vcov}
\alias{vcov,blm-method}
\alias{vcov,lexpit-method}

\title{Get variance-covariance from \code{blm} and \code{lexpit} objects.}

\description{

Extract large-sample Fisher Information approximation for covariance-variance matrix of the fit of a \code{blm} or \code{lexpit} model.

}

\section{Methods}{
\describe{

\item{vcov}{\code{signature(object = "blm")}:
Extractor for variance-covariance of MLEs. 
  }
 }
}


\author{S. Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
