% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_analysis.R
\name{bayes_factor}
\alias{bayes_factor}
\title{Bayes Factor}
\usage{
bayes_factor(model1, model2)
}
\arguments{
\item{model1}{first \code{blm} object}

\item{model2}{second \code{blm} object}
}
\description{
Computes the Bayes factor for comparison of 2 \code{\link{blm}} objects.
}
\details{
Returns the ratio of the probability of model1 divided by the
  probability of model2. A value of K > 1 means that M1 is more strongly
  supported by the data under consideration than M2
  (\url{https://en.wikipedia.org/wiki/Bayes_factor}). Wikipedia mentions
  additional scales for interpretation, ie by Harold Jeffreys:
   \itemize{
     \item K < 1: supports model2
     \item K > 10: strong support for model1
     \item K > 100: decisive support for model1
   }.
 An alternative table, widely cited, is provided by Kass and Raftery.
   \itemize{
     \item K 1-3: not worth more than a bare mention
     \item K 3-20: positive support for model1
     \item K 20-150: strong support for model1
     \item K >150: very strong support for model1
   }.
}
\examples{
set.seed(1) 
x <- seq(-10,10,.1) 
b <- 0.3

w0 <- 20 ; w1 <- 3 ; w2 <- 10

y <- rnorm(201, mean = w0 + w1 * x + w2 *sin(x), sd = sqrt(1/b)) 
model <- blm(y ~ x + sin(x), prior = NULL, data = data.frame(x=x, y=y))

plot(model, xlim=c(-10,10)) 
pblm(model) ## -955.8591 this is the log-probability

#another model removing the sinus term 
model2 <- blm(y ~ x, prior = NULL, data = data.frame(x=x, y=y))

pblm(model2) ## -958.128 
bayes_factor(model, model2) #9.668859

model3 <- blm(y ~ x + 0, prior = NULL, data = data.frame(x=x, y=y))
pblm(model3) ## -960.2644
bayes_factor(model, model3) #137.592

}

