% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{BayesLinReg-package}
\alias{BayesLinReg}
\alias{BayesLinReg-package}
\title{BayesLinReg package}
\description{
Bayesian linear regression from raw observations or sufficient statistics,
with normal, spike-and-slab, and global-local coefficient priors.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fmorgante/BayesLinReg}
  \item Report bugs at \url{https://github.com/fmorgante/BayesLinReg/issues}
}

}
\author{
\strong{Maintainer}: Fabio Morgante \email{fabiom@clemson.edu}

Authors:
\itemize{
  \item Fabio Morgante \email{fabiom@clemson.edu}
}

}
\keyword{internal}
