% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm_analysis.R
\name{bic}
\alias{bic}
\title{Bayesian information criterion (BIC)}
\usage{
bic(object, ...)
}
\arguments{
\item{object}{a \code{blm} object.}

\item{...}{other arguments (currently ignored).}
}
\description{
Bayesian information criterion (BIC) for the fit of a \code{blm} object.
}
\details{
Bayesian information criterion (BIC) of the \code{blm} object. 
  \eqn{BIC = n * ln(RSS/n) + k * ln(n)}. Where n is the number of 
  observations, k is the number of free parameters to be estimated and RSS is
  the residuals sum of squares (ie the \code{\link{deviance}}). When 
  comparing 2 models the strength of the evidence against the model with the 
  higher BIC value can be summarized as follows according to Wikipedia
  (\url{https://en.wikipedia.org/wiki/Bayesian_information_criterion}):
  \tabular{ll}{
    \bold{\eqn{\Delta}BIC} \tab \bold{Evidence against higher BIC}\cr
    0-2 \tab Not worth more than a bare mention\cr
    2-6 \tab Positive\cr
    6-10 \tab Strong\cr
    >10 \tab Very Strong\cr
  }
}
\examples{
set.seed(1) 
x <- seq(-10,10,.1) 
b <- 0.3

w0 <- 0.2 ; w1 <- 3 ; w2 <- 10

y <- rnorm(201, mean = w0 + w1 * x + w2 *sin(x), sd = sqrt(1/b)) 
model <- blm(y ~ x + sin(x), prior = NULL, data = data.frame(x=x, y=y))

plot(model, xlim=c(-10,10)) 
bic(model) ## 221.6777

#another model removing the sinus term 
model2 <- blm(y ~ x, prior = NULL, data = data.frame(x=x, y=y))

bic(model2) ## 805.4814
  
}

