\name{blm-package}
\alias{blm-package}
\docType{package}
\title{
Binomial linear and linear-expit regression  model
}
\description{
The functions \code{blm} and \code{lexpit} implement a binomial linear and linear-expit regression model. Estimates are the maximum likelihood estimates with constrained optimization through adaptive barrier method to ensure that estimable probabilities are in the (0,1) interval. 
}
\details{
\tabular{ll}{
Package: \tab blm\cr
Type: \tab Package\cr
Version: \tab 2013.2.4.2\cr
Date: \tab 2013-3-17\cr
Depends: R (>= 2.10.1), stats, methods, MASS
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes
}

}

\author{

Maintainer: Stephanie Kovalchik <s.a.kovalchik@gmail.com>

}

\seealso{\link{constrOptim}, \link{blm}, \link{lexpit}}

\keyword{ package }
