% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{blm}
\alias{blm}
\title{Bayesian linear model.}
\usage{
blm(model, ...)
}
\arguments{
\item{model}{a formula describing the model.}

\item{...}{additional data.}
}
\value{
a fitted model.
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}

