% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{blm}
\alias{blm}
\title{Create a blm object}
\usage{
blm(formula, data)
}
\arguments{
\item{formula}{linear regression formula}

\item{data}{data frame containing the variables in the linear regression formula}

\item{center}{TRUE if you want to grand-mean center the continuous variables in the data}
}
\value{
blm object containing roadmap for bayesian linear modeling
}
\description{
The blm() function sets up the bayesian linear model
}
