% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{blm}
\alias{blm}
\title{bayesian linear regression function}
\usage{
blm(formula, df, alpha = 0.5, beta = 1.3)
}
\arguments{
\item{formula}{an object of type formula}

\item{df}{an object of type dataframe}

\item{alpha}{a number}

\item{beta}{a number}
}
\value{
an object of class blm
}
\description{
main constructor function for creating an object of class blm
}

