% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{blm}
\alias{blm}
\title{bayesian linear regression function}
\usage{
blm(formula, beta = 1.3, prior = NULL, Data = NULL)
}
\arguments{
\item{formula}{an object of type formula}

\item{beta}{the precision}

\item{prior}{an object of type blm}

\item{Data}{an object of type dataframe}
}
\value{
an object of class blm
}
\description{
main constructor function for creating an object of class blm
}

