% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{blm}
\alias{blm}
\title{Bayesian Linear Regression}
\usage{
blm(formula, prior = NULL, beta = 1, ...)
}
\arguments{
\item{formula}{an object of class "formula": a symbolic description of the 
model to be fitted.}

\item{prior}{the prior distribution, either a \code{mv_dist} or a \code{blm} 
object.}

\item{beta}{the precision of the data used for the model.}

\item{...}{Additional arguments and values.}
}
\value{
A object of class \code{blm}. An object of class "lm" is a list 
  containing at least the following components: 
  \itemize{
    \item{call:      the matched call} 
    \item{formula:   the formula used} 
    \item{frame:     the model frame used} 
    \item{matrix:    the model matrix used} 
    \item{beta:      the precision of the data} 
    \item{prior:     the prior distribution used} 
    \item{posterior: the posterior distribution}
  }
}
\description{
Computes a bayesian linear regression fit. The current implementation 
requires knowledge about the precision \code{beta} of the observations.
}
\details{
Models for \code{blm} are provided as for \code{\link{lm}}. If prior
  distribution \code{prior} of the weights is not provided the prior means 
  are set to 0 and the variances to 1.
}
\examples{
  w0 <- 0.3 ; w1 <- 1.1 ; b <- 1.3 
  x <- rnorm(50)
  y <- rnorm(50, w1 * x + w0, 1/b) 
  mod <- blm(y~x, beta=b, data=data.frame(x=x, y=y))
  
  summary(mod) 
  
  plot(mod)
  
  #use of a prior, typically from an existing model 
  x2 <- rnorm(50) 
  y2 <- rnorm(50, w1 * x2 + w0, 1/b) 
  mod <- blm(y~x, prior=mod, beta=b, data=data.frame(x=x2, y=y2)) 
  mod
  
  #use with 2 explanatory variables 
  w2 <- 3.3 
  z <- rnorm(50) 
  y <- rnorm(50, w2 * z + w1 * x + w0, 1/b) 
  mod <- blm(y~x+z, beta=b, data=data.frame(x=x,y=y, z=z)) 
  mod
 
}

