% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\docType{package}
\name{blm}
\alias{blm}
\alias{blm-package}
\title{blm: A package for bayesian linear regression.}
\usage{
blm(formula, prior = NULL, beta = 1, ...)
}
\description{
The \code{blm} package is for construction and analysis of bayesian linear
 regression models.

Computes a bayesian linear regression fit. It only works for
 fixed variance beta.
 
 @param formula Object of class "formula": a symbolic description of the 
   model to be fitted.
 @param prior The prior distribution, either a \code{mv_dist} or a \code{blm}   object.
 @param beta Precision of the model fit.
 @param use_data_from_prior logical. If \code{TRUE} response and explanatory variables are extracted from the "prior". The prior needs to be provided as model object, ie \code{blm}.
 @param ... Additional arguments, or arguments with changed values
 
 @export
}
\section{blm functions}{

blm
update.blm
print.blm
summary.blm
resid.blm
coef.blm
deviance.blm
plot.blm
}

