% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{blm}
\alias{blm}
\title{Bayesian Linear Model}
\usage{
blm(model, prior, beta, ...)
}
\arguments{
\item{model}{A formula describing the model.}

\item{prior}{A prior distribution in the form of a list, containing at least a covariance matrix named sigma}

\item{beta}{The precision, a number}

\item{...}{Additional data, for example a data frame.}
}
\value{
A list of descriptive parameters of the prior and posterior distributions
}
\description{
Constructor for class 'blm'.
Calculates a posterior distribution of weights, given a model.
}

