% Generated by roxygen2 (4.0.2): do not edit by hand
\name{blm}
\alias{blm}
\title{Bayesian linear regression}
\usage{
blm(x, y, xp, yp, dfp = 1, scalepv)
}
\arguments{
\item{x}{The observed right-hand side variable data}

\item{y}{The observed left-hand side variable data}

\item{xp}{The dummy observation matrix for the right-hand side}

\item{yp}{The dummy observations for the left-hand side}

\item{dfp}{Degrees of freedom for the inverse-gamma prior on residual
variance}

\item{scalepv}{Scale for prior on variance}
}
\value{
List with elements
\itemize{
  \item lsout:  The returned output from \code{\link{lm.fit}}..
  \item postdf: Degrees of freedom in the inverse-gamma posterior for
        residual variance.
  \item postssr: Sum of squared residuals, over real and dummy observations,
        at posterior mode/mean.
  \item lmdd: log of marginal data density.
  \item vcv: Scale matrix for posterior \eqn{t} distribution on parameters.
             Variance matrix if multiplied by \code{postdf/(postdf-2)}
  \item sdcoeff: Vector of standard deviations of coefficients
        \code{sqrt(diag(vcv))}
}
}
\description{
Regression with a proper conjugate prior specified with dummy observations.
}
\details{
With dummy observations and degrees of freedom and scale for
the marginal distribution of disturbance variance, constructs
estimates and provides the marginal data density, allowing
model comparison via Bayes factors.

Constant term must be introduced explicitly as a column of ones in
   \code{x}.  With \code{dfp=1}, the inverse of residual variance has mean
   \code{1/scalevp}.  \code{scalevp} must be
   increased in proportion to \code{dfp} to keep the center of the distribution
   for residual variance stable as it becomes more precise.
}
\seealso{
\code{\link{tsregPrior}}, which prepares a symmetric prior, biased toward
   persistence, for models with lags.  \code{\link{lagts}} for preparing
   \code{mts} objects that can play the role of \code{x} in this program.
}

