% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridgeLinearModel.R
\name{blm}
\alias{blm}
\title{Bridge Linear Models}
\usage{
blm(
  data,
  intercept,
  weights,
  newFeatures,
  pThres = 1e-05,
  maxFeatPerUnknown = Inf,
  seed = 1
)
}
\arguments{
\item{data}{A matrix with named features in the rows, and unique sample names in the columns}

\item{intercept}{A single numeric value indicating the model intercept}

\item{weights}{A numeric vector corresponding the model feature weights.}

\item{newFeatures}{A character vector, indicating which features to use for the bridged model.}

\item{pThres}{TODO}

\item{maxFeatPerUnknown}{TODO}

\item{seed}{TODO}
}
\description{
By bridging linear models, a model can be described in terms of other features.
}
\details{
This function was developed specifically with the intention to convert a 
linear model known as SKY92 - which uses AffyMetrix probe IDs - into a model
that is described in terms of Ensembl Gene IDs. 

The AffyMetrix ProbeIDs reflect gene transcript expressions, while the Ensembl
gene IDs summarize the genes as a whole. Some genes might be targeted by
none, one or multiple probes, and similarly, some probes might correspond to
no, one or multiple Ensembl gene IDs. Although there might not be a one-to-one
relationship, there is most often a high correlation between features expressed
in terms of AffyMetrix Probe IDs and other representations such as Ensembl Gene
IDs. 

The function uses the underlying covariance structure in the data to 
redistribute the weight of a model over new features.
}
\examples{
set.seed(1)
#TODO!

}
\references{
Rowan Kuiper, Sonja Zweegman, Mark van Duin, Martin H. van Vliet, Erik H. van Beers, et.al; Prognostic and predictive performance of R-ISS with SKY92 in older patients with multiple myeloma: the HOVON-87/NMSG-18 trial. Blood Adv 2020; 4 (24): 6298–6309. doi:https://doi.org/10.1182/bloodadvances.2020002838
}
