% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{blm}
\alias{blm}
\title{Bayesian linear model.}
\usage{
blm(model, alpha, beta, data, ...)
}
\arguments{
\item{model}{A formula describing the model.}

\item{alpha}{Covariance from prior}

\item{beta}{Precision hyperparameter}

\item{data}{data for estimation, in data frame.}

\item{...}{Additional data}
}
\value{
An Object of class .blm, containing the inputted information as well as the fitted values of mean and sigma.
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}

