% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{blm}
\alias{blm}
\title{Bayesian linear model.}
\usage{
blm(model, prior, beta, ...)
}
\arguments{
\item{model}{A formula describing the model}

\item{prior}{A prior distribution}

\item{beta}{The precision}

\item{...}{Additional data.}
}
\value{
A fitted model.
}
\description{
The function fits a blm model, using a model formula, and optional additional data.
}

