% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.blm}
\alias{coef.blm}
\title{Coefficients}
\usage{
\method{coef}{blm}(object, ...)
}
\arguments{
\item{object}{The Bayesian linear regression model}

\item{...}{Additional data, for example a data frame.}
}
\value{
Fitted parameters for the model
}
\description{
Returns the coefficients of the Bayesian linear model
}

