% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{coef.blm}
\alias{coef.blm}
\title{Coefficients}
\usage{
\method{coef}{blm}(object, covar = FALSE, ...)
}
\arguments{
\item{object}{a \code{blm} object.}

\item{covar}{return the covariance matrix instead of estimates for the coefficients?}

\item{...}{other arguments (currently ignored).}
}
\value{
A named vector for the maxium a posteriori likelihood of coefficients. 
 That is, means of the posterior distribution of the coefficients of the
  object. To get the covariance matrix for the coefficients use 
  \code{\link{covar.blm}}.
}
\description{
Coefficients from the posterior distribution of a bayesian model.
}
\examples{
x <- rnorm(10)
b <- 1.3
w0 <- 0.2 ; w1 <- 3
y <- rnorm(10, mean = w0 + w1 * x, sd = sqrt(1/b))
model <- blm(y ~ x, prior = NULL, beta = b, data = data.frame(x=x, y=y))

coef(model)
coefficients(model)

}

