% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficients.R
\name{coef.blm}
\alias{coef.blm}
\title{Coefficients for the blm class}
\usage{
\method{coef}{blm}(x)
}
\arguments{
\item{x}{object of class blm}
}
\value{
vector of coefficients
}
\description{
Retrieve coefficients from the blm class
}

