% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMethods.R
\name{coef.blm}
\alias{coef.blm}
\title{Coefficients of Bayesian linear model.}
\usage{
\method{coef}{blm}(object, ...)
}
\arguments{
\item{object}{An object of class 'blm'}

\item{...}{Not used}
}
\value{
A list of coefficients from the posterior distributions
}
\description{
Reports the mean and covariance of the weights of a Bayesian linear model.
}

