% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMethods.R
\name{coef.blm}
\alias{coef.blm}
\title{The coefficients of a blm object}
\usage{
\method{coef}{blm}(object, ...)
}
\arguments{
\item{object}{the blm object}

\item{...}{any extra parameters}
}
\value{
the estimated parameters for the fitted model
}
\description{
This function calculates the coefficients of a blm object
}

