% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coef.R
\name{coef.blm}
\alias{coef.blm}
\title{Coefficients}
\usage{
\method{coef}{blm}(object, ...)
}
\arguments{
\item{object}{A (fitted) object of class blm}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\value{
a numeric vector with the coefficients from a fitted model
}
\description{
Retrieve parameters of the fitted model
}

