% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coef.blm}
\alias{coef.blm}
\title{Find coefficients for blm}
\usage{
\method{coef}{blm}(object, ...)
}
\arguments{
\item{object}{A blm model}

\item{...}{Extra parameters}
}
\value{
A list of coefficients
}
\description{
This function returns fitted parameters of the model.
}

