% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{coefficients.blm}
\alias{coefficients.blm}
\title{Bayesian linear model.}
\usage{
\method{coefficients}{blm}(x)
}
\arguments{
\item{model}{A formula describing the model.}
}
\value{
A fitted model.
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}

