% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.R
\name{coefficients.blm}
\alias{coefficients.blm}
\title{Bayesian linear model.}
\usage{
\method{coefficients}{blm}(object, ...)
}
\arguments{
\item{object}{a model object of class blm for which has at least one predictor variable.}

\item{...}{additional arguments.}
}
\value{
fitted arguments of the model (the point estimates of predictors).
}
\description{
coef is inspired in the generic function of stats and extracts model coefficients from objects .
}

