% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{coefficients.blm}
\alias{coefficients.blm}
\title{coefficients
A function for extracting the coefficients of a blm}
\usage{
\method{coefficients}{blm}(object)
}
\arguments{
\item{object}{an object of class blm}
}
\value{
a named vector of coefficients
}
\description{
coefficients
A function for extracting the coefficients of a blm
}

