% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient.R
\name{coefficients.blm}
\alias{coefficients.blm}
\title{Fitted Parameters of the model}
\usage{
\method{coefficients}{blm}(object)
}
\arguments{
\item{object}{A fit model obtained from blm function}
}
\value{
Fitted parameters of the model
}
\description{
This function returns the fitted parameters of the models
}

