% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{coefficients}
\alias{coefficients}
\title{Get coefficients of blm model.}
\usage{
coefficients(obj)
}
\arguments{
\item{obj}{A blm object.}
}
\value{
A named vector with coefficients.
}
\description{
Gets coefficients of a blm model.
}

