% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.blm}
\alias{confint.blm}
\title{Model confint
confint is inspired by stats function and computes the confidence intervals for the fitted arguments of a blm model.}
\usage{
\method{confint}{blm}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{a model object of class blm for which has at least one predictor variable.}

\item{parm}{a specified parameter/argument from the model to be extracted the confidence intervals. It could be numbers (columns of the model data.frame) or names: 'y', 'x', 'z'...}

\item{level}{the confidence level required, which quantiles are extracted in the base of: c(level/2, 1-level/2).}

\item{...}{additional arguments.}
}
\value{
It will returns the lower and upper confidence intervals of each parameter given.
}
\description{
Model confint
confint is inspired by stats function and computes the confidence intervals for the fitted arguments of a blm model.
}

