% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{confint.blm}
\alias{confint.blm}
\title{confidence interval function
A function for calculating confidence intervals of coefficients}
\usage{
\method{confint}{blm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{an object of class blm}

\item{parm}{a specification of which parameters are to be given confidence intervals}

\item{level}{a double/float}

\item{...}{further arguments passed to or from other methods}
}
\value{
confidence intervals of coefficients
}
\description{
confidence interval function
A function for calculating confidence intervals of coefficients
}

