% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confitblm.R
\name{confint.blm}
\alias{confint.blm}
\title{Confidence Intervals}
\usage{
\method{confint}{blm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{The Bayesian linear regression model}

\item{parm}{The parameter(s) in question}

\item{level}{The confidence level required, default gives 95\% CIs}

\item{...}{Additional data, for example a data frame.}
}
\value{
confidence intervals on fitted parameters
}
\description{
Calculates the confidence intervals for parameters fitted in the Bayesian linear model
}

