% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMethods.R
\name{confint.blm}
\alias{confint.blm}
\title{Confidence Interval for Bayesian Linear Model Weights}
\usage{
\method{confint}{blm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class 'blm'}

\item{parm}{A vector specifying which parameters to report CI's on.}

\item{level}{Confidence level. default = 0.95}

\item{...}{Not used}
}
\value{
A dataframe containing the weights and their upper and lower confidence limits.
}
\description{
Calculates the confidence interval of the weights
}

