% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMethods.R
\name{confint.blm}
\alias{confint.blm}
\title{The confidence intervals of a blm object}
\usage{
\method{confint}{blm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{the blm object}

\item{parm}{a vector of parameter names or a numeric vector with the wanted parameters}

\item{level}{the desired confidence level, default = 0.95}

\item{...}{any extra parameters}
}
\value{
a matrix with all the confidence intervals for the estimates
}
\description{
This function calculates the confidence intervals of all the estimates in a blm object
}

