% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.blm}
\alias{confint.blm}
\title{confint}
\usage{
\method{confint}{blm}(model, parameter, interval = 0.95)
}
\arguments{
\item{model}{A formula describing the model of class blm}

\item{parameter}{Parameter to calculate confidence interval for}

\item{interval}{Percentage of confidence interval limit, standard = 95%}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\value{
A confidence interval for parameter
}
\description{
Calculate confidence interval of fitted parameters
}

