% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.blm}
\alias{confint.blm}
\title{confint}
\usage{
\method{confint}{blm}(object, parm = 0, level = 0.95, ...)
}
\arguments{
\item{object}{An object of class blm}

\item{parm}{Parameter to calculate confidence interval for. Default 0 to return confidence interval for all weights}

\item{level}{Percentage of confidence interval limit, standard = 0.95}

\item{...}{Additional data}
}
\value{
A confidence interval for weights
}
\description{
Calculate confidence interval of fitted parameters
}

