% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.blm}
\alias{confint.blm}
\title{Find confidence intervals}
\usage{
\method{confint}{blm}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{A blm model}

\item{parm}{Number of parameters}

\item{level}{The confidence level}

\item{...}{Optional extra data}
}
\value{
Confidence intervals
}
\description{
This function finds the confidence intervals for the blm fitted parameters.
}

