% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{confint}
\alias{confint}
\title{Bayesian linear model.}
\usage{
confint(object, level = 0.95, ...)
}
\arguments{
\item{...}{Additional data, for example a data frame. Feel free to add other options.}

\item{model}{A formula describing the model.}
}
\value{
A fitted model.
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}

