% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{confint}
\alias{confint}
\title{Get confidence intervals.}
\usage{
confint(obj, parm, level = 0.95)
}
\arguments{
\item{obj}{A blm object.}

\item{parm}{Parameters for which CI has to be calculated.}

\item{level}{A parameter with CI level (0.95 for 95\% CI).}
}
\value{
A 2 rows matrix with confidence intervals.
}
\description{
Gets confidence intervals for coefficients from a fitted model.
}

