% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint}
\alias{confint}
\title{Confidence Interval for the fitted parameters}
\usage{
confint(object, level = 0.95)
}
\arguments{
\item{object}{A fiited model obtained from the blm}

\item{level}{A confidence level}
}
\value{
Confidence Interval for each of the parameters
}
\description{
This function gives the Confidence interval (2.5% and 97.5%) for the fitted parameters
}

