% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{construct_responseless_model_matrix}
\alias{construct_responseless_model_matrix}
\title{Construct responseless model matrix}
\usage{
construct_responseless_model_matrix(formula, df)
}
\arguments{
\item{formula}{an object of type formula}

\item{df}{an object of type dataframe}
}
\value{
a model matrix with data from df
}
\description{
This function constructs a model matrix for data with no response
}

