% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.blm}
\alias{deviance.blm}
\title{Model Deviance}
\usage{
\method{deviance}{blm}(object, ...)
}
\value{
The deviance value from the object of class blm.
}
\description{
It returns the deviance of a fitted blm model object.
}

