% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{deviance.blm}
\alias{deviance.blm}
\title{Deviance function
A function for calculating the deviance between fitted and observed values}
\usage{
\method{deviance}{blm}(object, ...)
}
\arguments{
\item{object}{an object of class blm}

\item{...}{further arguments passed to or from other methods}
}
\value{
the sum of squares of the model
}
\description{
Deviance function
A function for calculating the deviance between fitted and observed values
}

