% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.blm}
\alias{deviance.blm}
\title{Deviance}
\usage{
\method{deviance}{blm}(object, ...)
}
\arguments{
\item{object}{The Bayesian linear regression model}

\item{...}{Additional data, for example a data frame.}
}
\value{
deviance
}
\description{
Sum of the squared distances from the predicted response variable to the observed variable.
}

