% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{deviance.blm}
\alias{deviance.blm}
\title{Deviance of Bayesian Linear Model}
\usage{
\method{deviance}{blm}(object, ...)
}
\arguments{
\item{object}{a \code{blm} object.}

\item{...}{other arguments (currently ignored).}
}
\value{
Deviance of the \code{blm} object.
}
\description{
\code{deviance.blm} returns \code{\link{deviance}} for a bayesian linear
 model.
}
\details{
Deviance is the sum of squares of the residuals, extracted from the 
 maxium a posterior estimate of the \code{blm} object.
}
\examples{
x <- rnorm(10)
b <- 1.3
w0 <- 0.2 ; w1 <- 3
y <- rnorm(10, mean = w0 + w1 * x, sd = sqrt(1/b))
model <- blm(y ~ x, prior = NULL, beta = b, data = data.frame(x=x, y=y))

deviance(model)


}

