% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMethods.R
\name{deviance.blm}
\alias{deviance.blm}
\title{Deviance of Bayesian Linear Model Fit}
\usage{
\method{deviance}{blm}(object, ...)
}
\arguments{
\item{object}{An object of class 'blm'}

\item{...}{Not used}
}
\value{
A number
}
\description{
Reports the sum of squared residuals from a Bayesian linear model fit.
}

