% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMethods.R
\name{deviance.blm}
\alias{deviance.blm}
\title{The deviance of a blm object}
\usage{
\method{deviance}{blm}(object, ...)
}
\arguments{
\item{object}{the blm object}

\item{...}{any extra parameters}
}
\value{
the deviance
}
\description{
This function calculates the deviance of a blm object
}

