% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.blm}
\alias{deviance.blm}
\title{Deviance}
\usage{
\method{deviance}{blm}(object, ...)
}
\arguments{
\item{object}{A fiited model}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\value{
Deviance
}
\description{
This function just computes the sum of squared distances from the predicted response variables to the observed
}

