% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deviance.R
\name{deviance.blm}
\alias{deviance.blm}
\title{deviance}
\usage{
\method{deviance}{blm}(object, ...)
}
\arguments{
\item{object}{A formula describing the model of class blm}

\item{...}{Additifonal data, for example a data frame. Feel free to add other options.}
}
\value{
a numeric vector with sum of squared distances from the predicted response variables to the observed
}
\description{
sum of squared distances from the predicted response variables to the observed
}

