% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deviance.R
\name{deviance.blm}
\alias{deviance.blm}
\title{Calculate the deviance}
\usage{
\method{deviance}{blm}(object, ...)
}
\arguments{
\item{object}{A blm model}

\item{...}{extra data (not used)}
}
\value{
the deviance
}
\description{
This function finds the deviance for a blm.
}

