% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{deviance}
\alias{deviance}
\title{Get deviance.}
\usage{
deviance(obj)
}
\arguments{
\item{obj}{A blm object.}
}
\value{
A deviance.
}
\description{
Gets the sum of squared distances from the predicted to observed response.
}

