% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate_accepted_draws}
\alias{evaluate_accepted_draws}
\title{View the number of accepted posterior draws for each parameter and chain}
\usage{
evaluate_accepted_draws(x)
}
\arguments{
\item{x}{a blm object}
}
\value{
prints the number of accepted draws for each parameter in each chain to the console
}
\description{
If a user chooses to use the Gibbs sampler only, then the number of accepted draws is equal to the number of iterations in each chain. For the Metropolis-Hastings algorithm, this depends on the value of lambda set by the user in the \link[blm]{set_sampler} function.
}
