% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate_hypotheses}
\alias{evaluate_hypotheses}
\title{Evaluate hypotheses specified by the user}
\usage{
evaluate_hypotheses(x)
}
\arguments{
\item{x}{blm object}
}
\value{
returns the blm object containing Bayes' Factors
}
\description{
Evaluate hypotheses specified by the user
}
