% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_blm.R
\name{evaluate_model_fit}
\alias{evaluate_model_fit}
\alias{evaluate_model_fit.blm}
\title{Assess blm model fit}
\usage{
evaluate_model_fit(x)

\method{evaluate_model_fit}{blm}(x)
}
\arguments{
\item{x}{blm object}
}
\value{
prints summary of model fit (DIC) to the R console
}
\description{
Assess blm model fit
}
