% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate_ppc}
\alias{evaluate_ppc}
\title{Set up posterior predictive checks}
\usage{
evaluate_ppc(x, ...)
}
\arguments{
\item{x}{blm object}

\item{iterations}{number of iterations to run for posterior predictive checks. This number will be tagged on to the burn parameter specified under \link[blm]{set_sampling_options}.}

\item{return_samples}{logical. If TRUE, then the function will also return the posterior predictive samples.}
}
\value{
prints summary of the ppc to the R console
}
\description{
Set up posterior predictive checks
}
