% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{evaluate_R2}
\alias{evaluate_R2}
\title{Calculate the Bayesian R-squared value}
\usage{
evaluate_R2(x, iterations, return_samples)
}
\arguments{
\item{x}{a blm object}

\item{iterations}{number of samples to draw from the posterior distribution}

\item{return_samples}{logical. If TRUE, then the function will also return the posterior predictive samples.}
}
\value{
object of class "R2" containing input parameters, posterior draws and r-squared values computed on the posterior draws
}
\description{
Calculate the Bayesian R-squared value
}
\seealso{
Gelman, A., Goodrich, B., Gabry, J., & Vehtari, A. (2018). R-squared for Bayesian regression models. The American Statistician, (just-accepted), 1-6.
}
