% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.blm}
\alias{fitted.blm}
\title{Model fitted}
\usage{
\method{fitted}{blm}(object, ...)
}
\arguments{
\item{object}{a model object of class blm for which the extraction of model fitted values is desired.}

\item{...}{additional data.}
}
\value{
fitted values extracted from the object model of class blm.
}
\description{
fitted is inspired by the generic function from stats which extracts fitted values (response)
}
\seealso{
blm, fitted.blm, residuals.blm
}

