% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{fitted.blm}
\alias{fitted.blm}
\title{fitted values
A function for retrieving the fitted values of the response}
\usage{
\method{fitted}{blm}(object, ...)
}
\arguments{
\item{object}{an object of class blm}

\item{...}{further arguments passed to or from other methods}
}
\value{
a vector of fitted values
}
\description{
fitted values
A function for retrieving the fitted values of the response
}

