% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.blm}
\alias{fitted.blm}
\title{Fitted}
\usage{
\method{fitted}{blm}(object, ...)
}
\arguments{
\item{object}{The Bayesian linear regression model}

\item{...}{Additional data, for example a data frame (not used)}
}
\value{
Fitted response variables
}
\description{
Fitted response variable, fitted to the predictions that the Bayesian Linear Model makes
}

