% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMethods.R
\name{fitted.blm}
\alias{fitted.blm}
\title{Fitting a Bayesian linear model.}
\usage{
\method{fitted}{blm}(object, ...)
}
\arguments{
\item{object}{An object of class 'blm'}

\item{...}{Not used}
}
\value{
A data frame containing the predicted values and their variance.
}
\description{
Predicts response variable values from the data used to fit a Bayesian linear model.
}

