% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMethods.R
\name{fitted.blm}
\alias{fitted.blm}
\title{The fitted response of a blm object}
\usage{
\method{fitted}{blm}(object, ...)
}
\arguments{
\item{object}{the blm object}

\item{...}{any extra parameters}
}
\value{
the fitted responses
}
\description{
This function calculates the response the model given the data used to fit the model
}

