% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitted.R
\name{fitted.blm}
\alias{fitted.blm}
\title{Fitted Response Variables}
\usage{
\method{fitted}{blm}(object, ...)
}
\arguments{
\item{object}{A fitted model}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\value{
Fitted response variables
}
\description{
This function gives the fitted response variables.
}

