% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitted.R
\name{fitted.blm}
\alias{fitted.blm}
\title{fitted}
\usage{
\method{fitted}{blm}(object, ...)
}
\arguments{
\item{object}{An object class blm}

\item{...}{Additifonal data}
}
\value{
the fitted response variables.
}
\description{
This function should give you the fitted response variables.
}

