% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitted.R
\name{fitted.blm}
\alias{fitted.blm}
\title{Make a fitted model}
\usage{
\method{fitted}{blm}(object, ...)
}
\arguments{
\item{object}{A blm model}

\item{...}{Some extra data}
}
\value{
Fitted results
}
\description{
This function produces fitted response variables.
}

