% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fitted.R
\name{fitted}
\alias{fitted}
\title{Fitted Response Variables}
\usage{
fitted(object, ...)
}
\arguments{
\item{object}{A fitted model}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\value{
Fitted response variables
}
\description{
Takes a fitted model and optionally the data provided through the "..." parameter and returns the response variables
}

