% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_posterior_samples}
\alias{get_posterior_samples}
\title{Retrieve the posterior samples as a data frame from a blm object.}
\usage{
get_posterior_samples(x)
}
\arguments{
\item{x}{blm object}
}
\value{
a matrix of dimensions iterations x (variables + 1)
}
\description{
Retrieve the posterior samples as a data frame from a blm object.
}
