% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{get_value}
\alias{get_value}
\title{Get a value from an S3 object}
\usage{
get_value(x, var)
}
\arguments{
\item{x}{object with a get_value() method}

\item{var}{string. name of the value to retrieve from the object.}
}
\description{
Get a value from an S3 object
}
