\name{ccdata}
\docType{data}
\alias{ccdata}

\title{Simulated case-control dataset}

\description{Simulated population-based case-control dataset}
\usage{
ccdata
}
\format{
  \tabular{ll}{
    female:\tab indicator for female gender \cr
    packyear:\tab discrete variable representing pack-years smoked \cr
    strata: \tab stratification variable \cr
    y: \tab indicator of case status (1 for case, 0 for control) \cr
    w: \tab inverse of sampling fraction
  }
}

\keyword{datasets}
