% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior.R
\name{make_prior}
\alias{make_prior}
\title{Make the Prior Distribution}
\usage{
make_prior(model, alpha, ...)
}
\arguments{
\item{model}{A formula describing the model.}

\item{alpha}{The prior precision, the hyperparameter of Prior Distribution}

\item{...}{additional data}
}
\value{
The mean and covariance matrix
}
\description{
It takes the model description and the prior precision alpha to build the prior distribution
}

