% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior.R
\name{make_prior}
\alias{make_prior}
\title{Make prior.}
\usage{
make_prior(model, alpha, mu = c(0, 0))
}
\arguments{
\item{model}{A formula describing the model.}

\item{alpha}{A number}

\item{mu}{Mean, default set to 0}
}
\value{
A prior distribution
}
\description{
Makes a prior distribution for later use in Bayesian Linear Regression
}

