% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{make_prior}
\alias{make_prior}
\title{Make prior distribution.}
\usage{
make_prior(model, alpha, ...)
}
\arguments{
\item{model}{A formula describing the model.}

\item{alpha}{A parameter for prior distribution}

\item{...}{Additional data, for example a data frame.}
}
\value{
Prior distribution as a list with mu and Sigma..
}
\description{
Makes a prior distribution.
}

