% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{make_prior}
\alias{make_prior}
\title{Prior distribution}
\usage{
make_prior(mu, alpha, dim)
}
\value{
An object of class \code{mv_dist} containing a "means": vector of 
means and "covar": the covariance matrix.
}
\description{
Creates a prior distribution for a blm object
}

