% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{make_prior}
\alias{make_prior}
\title{Prior Distribution}
\usage{
make_prior(model, alpha)
}
\arguments{
\item{model}{A formula describing the model.}

\item{alpha}{Precision}
}
\value{
A list of the mean (always 0 for all parameters) and the covariance for the prior distribution.
}
\description{
Constructs a prior distribution, based on a known precision. Parameters are named such that they match 'blm' objects.
}

