% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Priors.R
\name{make_prior}
\alias{make_prior}
\title{Make the prior distribution}
\usage{
make_prior(alpha, n)
}
\arguments{
\item{alpha}{the alpha value}

\item{n}{the number of parameters in the model}
}
\value{
a list with a vector of means and a covariance matrix for the prior
}
\description{
This function makes a prior distribution given the alpha and number of parameters in the model
}

