% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prior.R
\name{make_prior}
\alias{make_prior}
\title{Prior Distribution}
\usage{
make_prior(model, alpha)
}
\arguments{
\item{model}{A formula describing the model}

\item{alpha}{A hyperparameter}
}
\value{
A list made of mean and sigma (the prior distribution)
}
\description{
Takes in a model and hyperparametr alpha and gives the prior distribution
}

