% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior.R
\name{make_prior}
\alias{make_prior}
\title{Making a prior.}
\usage{
make_prior(model, alpha, givenmean = c(0, 0))
}
\arguments{
\item{model}{A model function, fx. y ~ x + z}

\item{alpha}{A}

\item{givenmean}{....}
}
\value{
a prior distribution
}
\description{
This function produces a prior distribution.
}

