% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_matrix_repsonseless.R
\name{model_matrix_responseless}
\alias{model_matrix_responseless}
\title{Make a responseless matrix}
\usage{
model_matrix_responseless(model, data_frame)
}
\arguments{
\item{model}{A model function, fx. y ~ x + z}

\item{data_frame}{Some dataframe}
}
\value{
a prior distribution
}
\description{
This function removes the response variable from a matrix and returns a responseless matrix.
}

