% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution.R
\name{mv_dist}
\alias{mv_dist}
\title{Multivariate distribution}
\usage{
mv_dist(means, covar)
}
\arguments{
\item{means}{vector of means of the variables}

\item{covar}{covariance matrix of the variables}

\item{...}{currently ignored}
}
\value{
An object of class \code{distribution} containing: 
\describe{
   \item{means}{vector of means}
   \item{covar}{covariance matrix}
}
}
\description{
Means and covariance matrix of a multivariate normal distribution .
}

