% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blm}
\alias{plot.blm}
\title{Bayesian linear model.}
\usage{
\method{plot}{blm}(x, ...)
}
\arguments{
\item{x}{A formula describing the model.}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}

