% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot blm object}
\usage{
\method{plot}{blm}(object, ...)
}
\value{
plot of Residuals vs Fitted values.
}
\description{
It returns different plots of object blm
}
\seealso{
\code{\link[blm]}, \code{\link[blm]{fitted}}, \code{\link[blm]{residuals}}
}

