% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot blm x}
\usage{
\method{plot}{blm}(x, ...)
}
\arguments{
\item{x}{an blm x for which the plots are desired.}

\item{...}{additional arguments.}
}
\value{
plot of Residuals vs Fitted values.
}
\description{
It returns different plots of x blm
}
\seealso{
blm, fitted.
}

