% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot blm object}
\usage{
\method{plot}{blm}(object, ...)
}
\arguments{
\item{object}{an blm object for which the plots are desired.}

\item{...}{additional arguments.}
}
\value{
plot of Residuals vs Fitted values.
}
\description{
It returns different plots of object blm
}
\seealso{
blm, fitted.
}

