% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_blm.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot a blm object}
\usage{
\method{plot}{blm}(x, type = c("history", "autocorrelation", "density",
  "nullmodel"), ...)
}
\arguments{
\item{x}{blm object}

\item{type}{one of 'history' (trace plot), 'autocorrelation', 'density' or 'nullmodel'. The latter option plots all plots on one grid for the intercept-only model if it is computed. See also \code{compute_null_model()}}

\item{...}{other options passed for specific plots. Accepted arguments are:
\describe{
    \item{chain}{Integer >= 1 specifying which chain to use for autocorrelation plot. If not chosen, chain 1 is automatically used.}
}}
}
\description{
Plot a blm object
}
