% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{plot.blm}
\alias{plot.blm}
\title{plot function for blm
A function for plotting residuals versus fitted y values}
\usage{
\method{plot}{blm}(x, ...)
}
\arguments{
\item{x}{an object of class blm}

\item{...}{further arguments passed to or from other methods}
}
\value{
a plot of residuals versus fitted values
}
\description{
plot function for blm
A function for plotting residuals versus fitted y values
}

