% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot}
\usage{
\method{plot}{blm}(x, ...)
}
\arguments{
\item{x}{The Bayesian linear regression model}

\item{...}{Additional data, for example a data frame.}
}
\value{
plot
}
\description{
Plots the Bayesian linear model, the residuals vs the fitted values, similar to in a simple linear model
}

