% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot of a Bayesian Linear Model Fit}
\usage{
\method{plot}{blm}(object, explanatory = 1, ...)
}
\arguments{
\item{object}{a blm object.}

\item{explanatory}{a single number or vector for explanatory variables the 
response will be plotted against.}
}
\description{
Plots the blm object along with the MAP fit line and quantiles for the fit.
}

