% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMethods.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot of Bayesian linear model.}
\usage{
\method{plot}{blm}(x, ...)
}
\arguments{
\item{x}{An object of class 'blm'}

\item{...}{Not used}
}
\value{
A plot of residuals vs fitted values.
}
\description{
Plots the residuals vs the fitted values.
}

