% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMethods.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot a blm object}
\usage{
\method{plot}{blm}(x, param, ...)
}
\arguments{
\item{x}{the blm object}

\item{param}{a vector with the name of the responsevariable and the name of the predictor varialbe to be plotted. Onlye the first two variables in the vector is considered.}

\item{...}{any extra parameters}
}
\value{
a x-y plot with the data points and the line from the blm-fit and a residual plot
}
\description{
This function makes plots of a blm object.
}

