% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blm}
\alias{plot.blm}
\title{Plotting the variables}
\usage{
\method{plot}{blm}(x, ...)
}
\arguments{
\item{x}{A fitted model obtained from blm function}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\description{
This function plots the residuals against the fitted values
}

