% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot}
\usage{
\method{plot}{blm}(x, ...)
}
\arguments{
\item{x}{A fitted x of class blm}

\item{...}{Additifonal data, for example a data frame. Feel free to add other options.}
}
\value{
x-y plot with a line going through it for the fit
}
\description{
This function plots blm models
}

