% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.blm}
\alias{plot.blm}
\title{Plot a blm}
\usage{
\method{plot}{blm}(x, ...)
}
\arguments{
\item{x}{A blm model}

\item{...}{Some extra data (not used)}
}
\value{
a plot of the blm
}
\description{
This function produces a plot for a given blm.
}

