% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{plot}
\alias{plot}
\title{Make a plot.}
\usage{
plot(obj)
}
\arguments{
\item{obj}{A blm object.}
}
\value{
A plot.
}
\description{
Makes a plot with data points and a fitted line for models with one explanatory variable.
}

