% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{posterior}
\alias{posterior}
\title{get posterior of blm object}
\usage{
posterior(blm)
}
\arguments{
\item{blm}{an object of class blm}
}
\value{
a list of mean(s) and covariance matrix
}
\description{
This function fetches the mean(s) and covariance matrix posterior distribution
}

