% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.blm}
\alias{predict.blm}
\title{Model prediction}
\usage{
\method{predict}{blm}(object, ...)
}
\arguments{
\item{object}{a model object of class blm for which prediction is desired.}

\item{...}{additional data (explanatory variables) in data.frame format that affects the predictions. If ommited, the fitted values are used.}
}
\value{
If additional data is provided, it gives the predicted values by evaluating the regression function. Otherwise, it gives the predictions based on the data used to fit the model
}
\description{
Predict is inspired by the generic function from stats and make predictions based on the fitted blm model object.
}

