% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{predict.blm}
\alias{predict.blm}
\title{predict new values from blm}
\usage{
\method{predict}{blm}(blm, x)
}
\arguments{
\item{blm}{an object of class blm}

\item{x}{an object of type dataframe}
}
\value{
predicted y values from x
}
\description{
A function for predicting values of y from new data
}

