% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{predict.blm}
\alias{predict.blm}
\title{predict new values from blm}
\usage{
\method{predict}{blm}(object, newdata = NULL)
}
\arguments{
\item{object}{an object of class blm}

\item{newdata}{an object of type dataframe}
}
\value{
predicted y values from newdata
}
\description{
A function for predicting values of y from new data
}

