% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{predict.blm}
\alias{predict.blm}
\title{Predict}
\usage{
\method{predict}{blm}(object, newdata = NULL, report.var = F, ...)
}
\arguments{
\item{object}{a \code{blm} object.}

\item{newdata}{An optional data frame containing variables with which to 
predict. If missing, values used for fitting will be extracted from the 
object.}

\item{report.var}{Report also variance for the predicted values.}
}
\value{
A vector of predicted values. If report.var = TRUE, a list containing
 means and variances of the predicted values.
}
\description{
Predict means and variance of the response variable for a \code{\link{blm}}
 object.
}

