% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{predict.blm}
\alias{predict.blm}
\title{Predict}
\usage{
\method{predict}{blm}(object, newdata = NULL, se.fit = FALSE,
  interval = "none", level = 0.95, ...)
}
\arguments{
\item{object}{a \code{blm} object.}

\item{newdata}{an optional data frame containing variables with which to 
predict. If missing, values used for fitting will be extracted from the 
object.}

\item{se.fit}{report also standard deviation for the predicted values.}

\item{interval}{Type of interval calculation, currently only \code{none} and \code{confidence} implemented.}

\item{level}{confidence level for interval calculation.}

\item{...}{other arguments (currently ignored).}
}
\value{
A vector of predicted \code{fit} values. If \code{se.fit = TRUE}, a
  named list containing the fit values under \code{$fit} and standard
  deviations of the fit values under \code{$se.fit}. If \code{interval =
  "confidence"}, pseudo-confidence interval, ie lower and upper bounds of
  quantiles of the fit distrib ution are provided at \code{level} for the fit
  values are provided with \code{$fit} as data.frame with columns
  \code{$fit}, \code{$lwr}, \code{$upr}.
}
\description{
Predict means and variance of the response variable for a \code{\link{blm}} 
  object.
}
\examples{
  x <- rnorm(100)
  b <- 1.3
  w0 <- 0.2 ; w1 <- 3
  y <- rnorm(100, mean = w0 + w1 * x, sd = sqrt(1/b))
  model <- blm(y ~ x, prior = NULL, beta = b, data = data.frame(x=x, y=y))
  
  predict(model)
  
  #with standard deviation"of the fit distribution
  predict(model, se.fit=TRUE) 
  
  #with "confidence interval" of the fit values
  predict(model, interval = 'confidence', level = .95) 
  
  #predict for new explanatory values
  x <- rnorm(10) 
  predict(model, data.frame(x=x))

}

