% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.blm}
\alias{predict.blm}
\title{Predict new data from blm model}
\usage{
\method{predict}{blm}(object, ...)
}
\arguments{
\item{object}{Blm object}

\item{...}{Additional parameters such as data to be predicted on called 'data'. If left empty will return the fitted values. Boolean variable, which states whether variances of the predicted values should be included.}
}
\value{
predicted values
}
\description{
Predict new data points from a blm model, if no new data is given returns the fitted response variables. If new data is given, it gives predicted response variables.
}

