% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMethods.R
\name{predict.blm}
\alias{predict.blm}
\title{Predict response.}
\usage{
\method{predict}{blm}(object, newdata, ...)
}
\arguments{
\item{object}{An object of class 'blm'}

\item{newdata}{A data frame containing data different from that used to build the blm}

\item{...}{Not used}
}
\value{
see fitted().
}
\description{
Predict response of new data based on existing Bayesian linear model. It uses the old posterior distribution as a new prior distibution and then fits the new data by calling fitted().
}

