% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMethods.R
\name{predict.blm}
\alias{predict.blm}
\title{The predicted response of a blm object}
\usage{
\method{predict}{blm}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{the blm object}

\item{newdata}{the new data to predict response for, if not provided the response is predicted for the data used to fit the model}

\item{...}{any extra parameters}
}
\value{
the predicted responses
}
\description{
This function calculates the predicted response of a blm object
}

