% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict.blm}
\alias{predict.blm}
\title{Predict}
\usage{
\method{predict}{blm}(object, data = object$frame, ...)
}
\arguments{
\item{object}{A formula describing the model.}

\item{data}{Data frame containing data to predict from}

\item{...}{Additional data. Feel free to add other options.}
}
\value{
A mean vector
}
\description{
This function makes prediction based on the fitted model.
}

