% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Predict.R
\name{predict.blm}
\alias{predict.blm}
\title{Predict}
\usage{
\method{predict}{blm}(object, ...)
}
\arguments{
\item{object}{fitted model of type blm
If you do not provide new data, it instead gives you the predictions on the data used to fit the model.}

\item{...}{additional arguments affecting the predictions produced.}
}
\value{
predicted values based on fitted model
}
\description{
make predictions based on the fitted model
}

