% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.blm}
\alias{predict.blm}
\title{Make a prediction}
\usage{
\method{predict}{blm}(object, newdata, ...)
}
\arguments{
\item{object}{A blm model}

\item{newdata}{Data do be used}

\item{...}{Some extra inputs}
}
\value{
a prediction
}
\description{
This function makes a prediction for a blm, given some data.
}

