% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{predict}
\alias{predict}
\title{Predict response.}
\usage{
predict(obj, ...)
}
\arguments{
\item{obj}{A blm object.}

\item{...}{Additional data, newdata used to predict response.}
}
\value{
A vector with predicted response.
}
\description{
Predicts response based on fitted model, if no newdata is given, makes predictions on the data used to fit the model.
}

