% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.r
\name{predict}
\alias{predict}
\title{Predict}
\usage{
predict(object, ...)
}
\arguments{
\item{object}{A formula describing the model.}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\value{
A mean vector
}
\description{
This function makes prediction based on the fitted model.
}

