% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.blm}
\alias{print.blm}
\title{Model print}
\usage{
\method{print}{blm}(object, ...)
}
\value{
The linear function used and coefficients of an object of class "blm".
}
\description{
It prints the called function and the coefficients of a blm model object.
}

