% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{print.blm}
\alias{print.blm}
\title{print function for blm
A function for printing the call and coefficients of a blm}
\usage{
\method{print}{blm}(x, ...)
}
\arguments{
\item{x}{an object of class blm}

\item{...}{further arguments passed to or from other methods}
}
\value{
the function call and coefficients of the object
}
\description{
print function for blm
A function for printing the call and coefficients of a blm
}

