% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{print.blm}
\alias{print.blm}
\title{Print blm object.}
\usage{
\method{print}{blm}(x, ...)
}
\arguments{
\item{x}{A blm object.}

\item{...}{Additional parameters.}
}
\value{
Function call and coefficients.
}
\description{
Prints function call and coefficients of blm object.
}

