% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{print.blm}
\alias{print.blm}
\title{Print}
\usage{
\method{print}{blm}(x, ...)
}
\arguments{
\item{x}{a \code{\link{blm}} object.}

\item{...}{other arguments (currently ignored).}
}
\description{
Print a blm object
}

