% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMethods.R
\name{print.blm}
\alias{print.blm}
\title{Print a Bayesian linear model.}
\usage{
\method{print}{blm}(x, ...)
}
\arguments{
\item{x}{An object of class 'blm'}

\item{...}{Not used}
}
\value{
None.
}
\description{
Prints the function call that was made to fit the model as well as the coefficients (mean and covariance).
}

