% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMethods.R
\name{print.blm}
\alias{print.blm}
\title{Print a blm object}
\usage{
\method{print}{blm}(x, ...)
}
\arguments{
\item{x}{the blm object}

\item{...}{any extra parameters}
}
\value{
a print with the call and coefficients
}
\description{
This function prints the most general information from the object
}

