% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.blm}
\alias{print.blm}
\title{Print}
\usage{
\method{print}{blm}(x, ...)
}
\arguments{
\item{x}{A fit model obtained from blm function}

\item{...}{Additional data}
}
\description{
This function prints how the fitting function was called and what the fitted coefficients are.
}

