% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Print.R
\name{print.blm}
\alias{print.blm}
\title{Print}
\usage{
\method{print}{blm}(x, ...)
}
\arguments{
\item{x}{object of class .blm}

\item{...}{Additifonal data, for example a data frame. Feel free to add other options.}
}
\value{
prints a very short description of the object.
}
\description{
This function prints an object
}

