% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.blm}
\alias{print.blm}
\title{Print.}
\usage{
\method{print}{blm}(x, ...)
}
\arguments{
\item{x}{A blm model.}

\item{...}{Additional data.}
}
\value{
A printed summary of the blm object.
}
\description{
Gives a printed summary of the parameters of a blm model.
}

