% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.blm}
\alias{print.summary.blm}
\title{blm Summary}
\usage{
\method{print}{summary.blm}(object, ...)
}
\arguments{
\item{object}{an blm object for which summary is desired.}

\item{...}{additional arguments.}
}
\value{
MAKE HERE IN MORE DETAIL.
}
\description{
It is inpired by the generic and lm function 'summary' and it aims to format and summarize the statistical outputs from an object of class 'blm'.
}

