% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{print.summary.blm}
\alias{print.summary.blm}
\title{Print}
\usage{
\method{print}{summary.blm}(x, ...)
}
\arguments{
\item{x}{a \code{\link{summary.blm}} object.}

\item{...}{other arguments (currently ignored).}
}
\description{
Print a summary.blm object
}

