\name{resid}
\docType{methods}
\alias{resid}
\alias{resid,blm-method}
\alias{resid,lexpit-method}

\title{Get residuals from \code{blm} and \code{lexpit} objects.}

\description{
Extract residuals of model fit.
}

\section{Methods}{
\describe{

\item{resid}{\code{signature(object = "blm")}:
Extractor for residuals of \code{blm} object. 
  }

\item{resid}{\code{signature(object = "lexpit")}:
Extractor for residuals of \code{blm} object. 
  }

 }
}


\author{Stephanie Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
