% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.blm}
\alias{residuals.blm}
\title{Model Residuals}
\usage{
\method{residuals}{blm}(object, ...)
}
\value{
residuals extracted from the blm object.
}
\description{
It returns the model residuals from a blm object.
}
\seealso{
coefficients.blm, fitted.blm, blm, predict.blm
}

