% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{residuals.blm}
\alias{residuals.blm}
\title{residual function
A function for calculating residuals}
\usage{
\method{residuals}{blm}(object, ...)
}
\arguments{
\item{object}{an object of class blm}

\item{...}{further arguments passed to or from other methods}
}
\value{
a vector of residuals
}
\description{
residual function
A function for calculating residuals
}

