% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.blm}
\alias{residuals.blm}
\title{Residuals}
\usage{
\method{residuals}{blm}(object, ...)
}
\arguments{
\item{object}{The Bayesian linear regression model}

\item{...}{Additional data, for example a data frame.}
}
\value{
residuals
}
\description{
Returns the residual fitted Bayesian model, the difference between the predicted and the observed values.
}

