% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{residuals.blm}
\alias{residuals.blm}
\title{Residuals of Bayesian Linear Model}
\usage{
\method{residuals}{blm}(object, var = FALSE, ...)
}
\arguments{
\item{object}{a \code{blm} object.}

\item{var}{Report also variance for the fitted values.}

\item{...}{other arguments (currently ignored).}
}
\value{
Residuals of the \code{blm} object.
}
\description{
\code{residuals.blm} returns \code{\link{residuals}} for a bayesian linear
 model.
}
\details{
Residuals are extracted from the maxium a posterior estimate of the 
  \code{blm} object. If var = TRUE, a list containing means and variances of
  the fitted values.
}
\examples{
  x <- rnorm(10)
  b <- 1.3
  w0 <- 0.2 ; w1 <- 3
  y <- rnorm(10, mean = w0 + w1 * x, sd = sqrt(1/b))
  model <- blm(y ~ x, prior = NULL, beta = b, data = data.frame(x=x, y=y))
  
  residuals(model)
  resid(model)
  
}

