% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMethods.R
\name{residuals.blm}
\alias{residuals.blm}
\title{Residuals from Bayesian Linear Model Fit}
\usage{
\method{residuals}{blm}(object, ...)
}
\arguments{
\item{object}{An object of class 'blm'}

\item{...}{Not used}
}
\value{
A vector of residuals.
}
\description{
Calculates the difference between observed and predicted data.
}

