% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMethods.R
\name{residuals.blm}
\alias{residuals.blm}
\title{The residuals of a blm object}
\usage{
\method{residuals}{blm}(object, ...)
}
\arguments{
\item{object}{the blm object}

\item{...}{any extra parameters}
}
\value{
the residuals for the fitted data
}
\description{
This function calculates the residuals of a blm object
}

