% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residuals.R
\name{residuals.blm}
\alias{residuals.blm}
\title{Residuals}
\usage{
\method{residuals}{blm}(object, ...)
}
\arguments{
\item{object}{object of class .blm}

\item{...}{Additifonal data, for example a data frame. Feel free to add other options.}
}
\value{
the difference between predicted values and observed values for the response variable.
}
\description{
This function returns the residuals of the fit.
}
\details{
That is the difference between predicted values and observed values for the response variable.
}

