% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.blm}
\alias{residuals.blm}
\title{Find the residuals}
\usage{
\method{residuals}{blm}(object, ...)
}
\arguments{
\item{object}{A blm model}

\item{...}{some additional data}
}
\value{
The residuals
}
\description{
This function returns the residuals of a blm.
}

