% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{residuals}
\alias{residuals}
\title{Get residuals.}
\usage{
residuals(obj)
}
\arguments{
\item{obj}{A blm object.}
}
\value{
A vector with residuals.
}
\description{
Gets residuals of the fit; the difference between predicted and observed values of the response.
}

