% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Residuals.R
\name{residuals}
\alias{residuals}
\title{Residuals}
\usage{
residuals(object, ...)
}
\arguments{
\item{object}{A fitted model}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\value{
Residuals of the fit
}
\description{
This function returns the residuals of the fit.
}

