% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{sample_posterior}
\alias{sample_posterior}
\title{Sample the posterior distribution}
\usage{
sample_posterior(x)
}
\arguments{
\item{x}{blm object}
}
\value{
blm object containing sampled posterior
}
\description{
Sample the posterior distribution
}
