% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_blm.R
\name{set_sampler}
\alias{set_sampler}
\alias{set_sampler.blm}
\title{Update the sampler used for a specific variable}
\usage{
set_sampler(x, ...)

\method{set_sampler}{blm}(x, par, type = c("Gibbs", "MH"))
}
\arguments{
\item{x}{a blm object}

\item{...}{other arguments passed to the function}

\item{par}{name of the coefficient, given as b0 (intercept) or b1, ... for coefficients.}

\item{type}{name of the sampler to use. Either 'Gibbs' or 'MH'. Defaults to 'Gibbs'.}
}
\value{
updated blm object
}
\description{
The use can set the sampler to either (1) 'Gibbs' (default for all coefficients) or (2) 'MH' (metropolis-hastings algorithm, only for coefficients).
}
\examples{
bfit <- set_sampler(bfit, "b0", type="MH")

}
