% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/methods_blm.R
\name{set_sampling_options}
\alias{set_sampling_options}
\alias{set_sampling_options.blm}
\title{Set sampling options on a blm object}
\usage{
set_sampling_options(x, ...)

\method{set_sampling_options}{blm}(x, chains = 1, iterations = 10000,
  thinning = 1, burn = 1000)
}
\arguments{
\item{x}{blm object}

\item{...}{other options (chains, iterations etc.)}

\item{chains}{number of chains to initialize}

\item{iterations}{number of iterations to run the MC sampler}

\item{thinning}{set thinning parameter}

\item{burn}{number of samples that will be discarded}
}
\value{
updated blm object
}
\description{
Set sampling options on a blm object
}
\examples{
bfit <- set_sampling_options(bfit, chains=2, iterations=20000, thinning=3, burn = 2000)

}
