\name{show}
\docType{methods}
\alias{show}
\alias{show,blm-method}
\alias{show,lexpit-method}

\title{Show \code{blm} and \code{lexpit} model fit.}

\description{

Print estimates of a \code{blm} or \code{lexpit} model fit.

}

\section{Methods}{
\describe{
\item{show}{\code{signature(object = "blm")}:
Call, coefficient estimates, and AIC for fitted model.
  }

\item{show}{\code{signature(object = "lexpit")}:
Call, coefficient estimates, and AIC for fitted model.
  }
  }
 }

\author{Stephanie Kovalchik \email{s.a.kovalchik@gmail.com}}

\keyword{methods}
