% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.blm}
\alias{summary.blm}
\title{Summary}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{The Bayesian linear regression model}

\item{...}{Additional data, for example a data frame.}
}
\value{
summary
}
\description{
Summarises the features of the Bayesian regression model
}

