% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{summary.blm}
\alias{summary.blm}
\title{Print blm object.}
\usage{
\method{summary}{blm}(obj)
}
\arguments{
\item{obj}{A blm object.}
}
\value{
Function call and coefficients.
}
\description{
Prints function call and coefficients of blm object.
}

