% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{summary.blm}
\alias{summary.blm}
\title{Summary of blm object.}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{A blm object.}

\item{...}{Additional parameters.}
}
\value{
Function call and coefficients.
}
\description{
Prints summary of blm object.
}

