% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{summary.blm}
\alias{summary.blm}
\title{Summary}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{a \code{\link{blm}} object.}

\item{...}{other arguments (currently ignored).}
}
\value{
An object of class \code{summary.blm} containing:
 \describe{
  \item{call}{the call of the object }
  \item{prior}{the prior distribution of the fit}
  \item{posterior}{the posterior distribution of the fit}
 }
}
\description{
Summary of a blm object
}

