% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelMethods.R
\name{summary.blm}
\alias{summary.blm}
\title{Summary of Bayesian linear model.}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{An object of class 'blm'}

\item{...}{Not used}
}
\value{
None.
}
\description{
Prints summary statistics for a blm fit: Residuals, Deviance.
}

