% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelMethods.R
\name{summary.blm}
\alias{summary.blm}
\title{Print a summary of a blm object}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{the blm object}

\item{...}{any extra parameters}
}
\value{
a print with the call, a summary of the residuals, the coefficients, the residual standard error, the R-squared and adjusted R-squared
}
\description{
This function prints the more extensive information from the object
}

