% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.blm}
\alias{summary.blm}
\title{Summary}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{A fit model obtained from blm function}

\item{...}{optional parameters}
}
\description{
This function prints how the fitting function was called, the fitted coefficients and the residuals of the fit.
}

