% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{summary.blm}
\alias{summary.blm}
\title{Summary}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{fitted model of class blm}

\item{...}{Additifonal data}
}
\value{
returns an object with summary information
}
\description{
This function returns an oject with a summary of the fitted model
}

