% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{summary.blm}
\alias{summary.blm}
\title{Summary}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{...}{Additifonal data, for example a data frame. Feel free to add other options.}

\item{model}{fitted model of class blm}
}
\value{
returns an object with summary information
}
\description{
This function returns an oject with a summary of the fitted model
}

