% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.blm}
\alias{summary.blm}
\title{Summary.}
\usage{
\method{summary}{blm}(object, ...)
}
\arguments{
\item{object}{A blm object.}

\item{...}{Additional data.}
}
\value{
A summary of a blm object.
}
\description{
Gives a summary of a blm regression model, using the print.blm function.
}

