% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Bayesian linear model.}
\usage{
summary(x, ...)
}
\arguments{
\item{x}{A formula describing the model.}

\item{...}{Additional data, for example a data frame. Feel free to add other options.}
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}

