% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{Summary}
\alias{Summary}
\title{Summary}
\usage{
Summary(object)
}
\arguments{
\item{object}{A fit model obtained from blm function}
}
\description{
This function prints how the fitting function was called, the fitted coefficients and the residuals of the fit.
}

