% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Bayesian linear model.}
\usage{
update(model, prior, alpha, beta, ...)
}
\arguments{
\item{model}{A formula describing the model.}

\item{prior}{The prior distribution (mean and variance)}

\item{alpha}{Hyperparameter of prior distribution.}

\item{beta}{Hyperparameter of posterior distribution.}

\item{...}{Additional data, for example a data frame.}
}
\value{
A fitted model.
}
\description{
Fits a model, given as a formula, optionally with data provided through the "..." parameter.
}

