% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blm.R
\name{update}
\alias{update}
\title{Update prior to posterior distribution.}
\usage{
update(model, prior, beta, ...)
}
\arguments{
\item{model}{A formula describing the model.}

\item{prior}{Prior distribution}

\item{beta}{A parameter for posterior distribution}

\item{...}{Additional data, for example a data frame.}
}
\value{
Posterior distribution as a list with mu and Sigma.
}
\description{
Updates posterior distribution based on new prior/data
}

