% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supportMethods.R
\name{update}
\alias{update}
\title{Posterior Distribution}
\usage{
update(model, prior, beta, ...)
}
\arguments{
\item{model}{A formula describing the model.}

\item{prior}{A prior distribution, must have at least a covariance matrix named sigma}

\item{beta}{The precision}

\item{...}{Additional data, for example a data frame.}
}
\value{
A list contating the mean and the covariance matrix for the posterior distribution as well as the original data packed into a data frame.
}
\description{
Updates the prior distribution using data to give the posterior distribution og weights.
Support method for class 'blm'
}

