% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Posterior Distribution}
\usage{
update(model, prior, beta, ...)
}
\arguments{
\item{model}{A formula describing the model}

\item{prior}{Prior Distribution}

\item{beta}{A precision parameter}

\item{...}{Additional data}
}
\value{
A list made of mean and sigma (the prior distribution)
}
\description{
Takes in a model, prior distribution and the precision parameter and returns the updated posterior distribution
}

