% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\title{Update function.}
\usage{
update(model, prior, beta, ...)
}
\arguments{
\item{model}{A model function, fx. y ~ x + z}

\item{prior}{A prior distribution including a Sigma}

\item{beta}{The precision}

\item{...}{Additional data}
}
\value{
A list with Sigma, mean and data for a posterior distribution
}
\description{
This function updates a prior distribution, to get the posterior distribution.
}

