% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{update_posterior}
\alias{update_posterior}
\title{Draw more samples from a sampled posterior distribution.}
\usage{
update_posterior(x, ...)
}
\arguments{
\item{x}{blm object}

\item{iterations}{number of additional samples to draw from the posterior.}
}
\value{
blm object with updated count of posterior samples.
}
\description{
Draw more samples from a sampled posterior distribution.
}
\examples{
data("directors")
fit <- blm("Compensation ~ Age", data=directors) \%>\%
   sample_posterior()
# Update posterior samples
fit <- fit \%>\% update_posterior(iterations=2000)

}
